### 2020-10-01 #2713 @gaelforget
Hi @Kevin-Mattheus-Moerman

Thanks for taking this on and for your amazingly quick first assessment. I am especially grateful for such efficiency in light of the strain that COVID has been putting on all of us and the fact that joss is operating in reduced service mode in particular.

Will get back to you with specifics asap after I re-read the journal docs another time in light of your comment -- I feel I may have overlooked something subtle in the language of the submission guidelines as I was under the impression that this submission would rather easily clear the bar for getting to the open review stage.

In any event, I look forward to additional feedback that you or any other editor may post here, will do my best to respond as rapidly as I can, and hope that I will get a chance to address possible concerns about e.g. functionality or scholarly work over the course of the review process if needed.

By the way, is it ok for me to ask follow up questions here if that's going to help me respond to your comment in more detail?
