### 2020-10-08 #2713 @gaelforget
Hi @Kevin-Mattheus-Moerman

A quick post to apologize for the delay in my response -- it's been a rather hectic week on on my end ...

If that's ok, let me start with

> clearly and concisely outline the different types of functionality this work offers

and then turn to the scope and scholarly effort aspects later today and tomorrow
