### 2020-10-08 #2713 @gaelforget
Below I outline the different types of functionality this work offers.

The [three_dimensional_ocean notebook](https://juliaclimate.github.io/IndividualDisplacements.jl/dev/worldwide/three_dimensional_ocean/) and [the corresponding animation](https://youtu.be/twAAE_WUs_g) additionally provide a representative example of these functionalities put to action.

[W.I.P.]
