### 2020-10-09 #2713 @kthyng
@gaelforget thanks for the extra information! We can move forward with this submission. I will be your editor since I do particle tracking too, though not with Julia. I will need particular help from you finding reviewers who can review this software in Julia. Please provide github handles or another way to reach people from either your own network (but not conflicting) or from the list [of volunteers](https://bit.ly/joss-reviewers). Do not use "@" when mentioning github handles here so we don't spam people unnecessarily.
