### 2020-10-14 #2713 @gaelforget
> @gaelforget thanks for the extra information! We can move forward with this submission. I will be your editor since I do particle tracking too, though not with Julia. I will need particular help from you finding reviewers who can review this software in Julia. Please provide github handles or another way to reach people from either your own network (but not conflicting) or from the list [of volunteers](https://bit.ly/joss-reviewers). Do not use "@" when mentioning github handles here so we don't spam people unnecessarily.

This is great news! Thanks @kthyng for taking on this submission as editor.

I will work on reviewer suggestions asap and hope to get back to you later today on this. Thanks for your patience
