### 2020-10-19 #2713 @kthyng
@gaelforget Can you add a short section to your paper also to compare your software with other particle tracking software? This will help give context.
