### 2020-10-27 #2713 @gaelforget
> @gaelforget Still waiting to hear back from you about reviewer suggestions too.

Based on the list of volunteers (Preferred Language + Domain Expertise, mostly) and from my own network (incl. https://github.com/JuliaClimate) I may suggest:

- juliohm, felixcremer, visr, Alexander-Barth, natgeo-wong, milankl
- wkearn, Jgoldfar, eviatarbach (not in JuliaClimate)

Would this work?

ps. Others that came to mind but I felt may be too closely connected to me (`conflicting`?) include: ali-ramadhan and hdrake who are in my department at MIT; the CLIMA project which involves many close collaborators although not myself.
