### 2020-10-30 #2713 @kthyng
Hi @visr and @milankl! Are you interested in reviewing this submission to the Journal for Open Source Software? The review is done openly through github and is checklist-driven. Acknowledging the global pandemic, we're asking for reviews in 5-6 weeks. You can get more information from our [docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) (though looks like they are offline at the moment). Please feel free to ask questions too.

I encourage you to [sign up](https://joss.theoj.org/reviewer-signup.html) as a JOSS reviewer if you'd like!
