### 2020-10-31 1328630
use OceanStateEstimation.jl Artifacts folders for input
Former-commit-id: 4fc087ef17d117b5729ce3033de512d82336f576
