+++ examples/helper_functions.jl
@@ -17,18 +17,12 @@ end
 Download `MITgcm` transport output to `examples/nctiles_climatology` if needed
 """
 function get_ecco_velocity_if_needed()
-    p=dirname(pathof(IndividualDisplacements))
-    q=dirname(pathof(OceanStateEstimation))
-    pth0=pwd()
-    cd(joinpath(p,"../examples/"))
-    lst="nctiles_climatology.csv"
-    pth="nctiles_climatology/"
-    !isfile(lst) ? run(`cp $q/../examples/nctiles_climatology.csv $p/../examples/`) : nothing
-    !isdir("$pth") ? mkdir("$pth") : nothing
-    !isdir("$pth"*"UVELMASS") ? get_from_dataverse(lst,"UVELMASS",pth) : nothing
-    !isdir("$pth"*"VVELMASS") ? get_from_dataverse(lst,"VVELMASS",pth) : nothing
-    !isdir("$pth"*"WVELMASS") ? get_from_dataverse(lst,"WVELMASS",pth) : nothing
-    cd(pth0)
+    p=dirname(pathof(OceanStateEstimation))
+    lst=joinpath(p,"../examples/nctiles_climatology.csv")
+    pth=ECCOclim_path
+    !isdir(pth*"UVELMASS") ? get_from_dataverse(lst,"UVELMASS",pth) : nothing
+    !isdir(pth*"VVELMASS") ? get_from_dataverse(lst,"VVELMASS",pth) : nothing
+    !isdir(pth*"WVELMASS") ? get_from_dataverse(lst,"WVELMASS",pth) : nothing
 end
 
 """
@@ -49,22 +43,16 @@ end
 Download `MITgcm` transport output to `examples/OCCA_climatology` if needed
 """
 function get_occa_velocity_if_needed()
-    p=dirname(pathof(IndividualDisplacements))
-    q=dirname(pathof(OceanStateEstimation))
-    pth0=pwd()
-    cd(joinpath(p,"../examples/"))
-    lst="OCCA_climatology.csv"
-    pth="OCCA_climatology/"
-    !isfile(lst) ? run(`cp $q/../examples/OCCA_climatology.csv $p/../examples/`) : nothing
+    p=dirname(pathof(OceanStateEstimation))
+    lst=joinpath(p,"../examples/OCCA_climatology.csv")
+    pth=OCCAclim_path
     nams = ("DDuvel.0406clim.nc","DDvvel.0406clim.nc","DDwvel.0406clim.nc","DDtheta.0406clim.nc","DDsalt.0406clim.nc")
-    !isdir("$pth") ? mkdir("$pth") : nothing
     if !isfile("$pth"*"DDuvel.0406clim.nc") 
-        tmp=joinpath(dirname(pathof(IndividualDisplacements)),"../examples/OCCA_climatology/tmp/")
-        !isdir("$tmp") ? mkdir("$tmp") : nothing
+        tmp=joinpath(pth,"tmp/")
+        !isdir(tmp) ? mkdir(tmp) : nothing
         [get_from_dataverse(lst,nam,tmp) for nam in nams]
         [mv(joinpath(tmp,nam,nam),joinpath(pth,nam)) for nam in nams]
     end
-    cd(pth0)
 end
 
 """
@@ -128,7 +116,7 @@ function setup_global_ocean(;k=1,ny=2)
[…]
