+++ .travis.yml
@@ -2,7 +2,7 @@ language: julia
 os:
   - linux
 julia:
-  - 1.2
+  - 1.3
   - 1.5
   - nightly
 notifications:
@@ -19,7 +19,7 @@ after_success:
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.2
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
