+++ .travis.yml
@@ -24,7 +24,7 @@ jobs:
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
                                     Pkg.develop(PackageSpec(path=pwd()))'
-        - julia --project=docs/ docs/make.jl
+        - travis_wait 30 julia --project=docs/ docs/make.jl
       after_success: skip
   allow_failures:
     - julia: nightly
