### 2020-11-01 7ff4c40
get grid files from MeshArrays Artifacts
Former-commit-id: 9acd4363e453c466e5c9e231e54e81982dae4804
