### 2020-11-01 9acd436
get grid files from MeshArrays Artifacts
