### 2020-11-01 dd35f37
reduce example problem sizes
Former-commit-id: f8fd6265e5bb6a4d8573f51913a51234ec93c45a
