+++ Project.toml
@@ -24,7 +24,7 @@ CSV = "0.6, 0.7"
 CyclicArrays = "0.2, 0.3, 0.4"
 DataFrames = "0.21"
 MITgcmTools = "0.1.6"
-MeshArrays = "0.2.13"
+MeshArrays = "0.2.14"
 NetCDF = "0.10"
 OceanStateEstimation = "0.1.5"
 OrdinaryDiffEq = "5"
