### 2020-11-01 a545970
fix case of empty lst2
Former-commit-id: 012ab3548a7f1a2cdfa8a17a2345597dc8cd4f2b
