+++ docs/make.jl
@@ -21,8 +21,8 @@ nb = true      # Set to true to generate the notebooks
 
 lst1 = ["solid_body_rotation","random_flow_field","global_ocean_circulation","three_dimensional_ocean","detailed_look","particle_cloud"]
 lst2 = []
-tst1(x) = Bool(sum(isequal.(x, lst1)))
-tst2(x) = Bool(sum(isequal.(x, lst2)))
+tst1(x) = !isempty(lst1) && Bool(sum(isequal.(x, lst1)))
+tst2(x) = !isempty(lst2) && Bool(sum(isequal.(x, lst2)))
 
 for (root, _, files) in walkdir(lit), file in files
     splitext(file)[2] == ".jl" || continue
