### 2020-11-01 ce43959
fix case of empty lst2 (?)
