+++ docs/make.jl
@@ -20,7 +20,7 @@ execute = true # Set to true for executing notebooks and documenter!
 nb = true      # Set to true to generate the notebooks
 
 lst1 = ["solid_body_rotation","random_flow_field","global_ocean_circulation","three_dimensional_ocean","detailed_look","particle_cloud"]
-lst2 = []
+lst2 = ["none"]
 tst1(x) = !isempty(lst1) && Bool(sum(isequal.(x, lst1)))
 tst2(x) = !isempty(lst2) && Bool(sum(isequal.(x, lst2)))
 
