### 2020-11-02 a41d433
try skipping jldoctest in compute
