+++ src/compute.jl
@@ -4,7 +4,7 @@
 Interpolate velocity from gridded fields (2D; with halos) to position `u`
 (`x,y,fIndex`) to compute the derivative of position v time  `du_dt`.
 
-```jldoctest
+```gfdoctest
 using IndividualDisplacements, Statistics
 p=dirname(pathof(IndividualDisplacements))
 include(joinpath(p,"../examples/basics/random_flow_field.jl"))
@@ -16,7 +16,7 @@ prod(isapprox.([mean(𝐼.🔴.x) mean(𝐼.🔴.y)],ref,atol=10.0))
 true
 ```
 
-```jldoctest
+```gfdoctest
 using IndividualDisplacements, Statistics
 p=dirname(pathof(IndividualDisplacements))
 include(joinpath(p,"../examples/worldwide/global_ocean_circulation.jl"))
@@ -81,7 +81,7 @@ end
 Interpolate velocity from gridded fields (3D; NO halos) to position `u`
 (`x,y,z`) to compute the derivative of position v time  `du_dt`.
 
-```jldoctest
+```gfdoctest
 using IndividualDisplacements
 p=dirname(pathof(IndividualDisplacements))
 include(joinpath(p,"../examples/basics/solid_body_rotation.jl"))
@@ -93,7 +93,7 @@ prod(isapprox.(𝐼.📌',ref))
 true
 ```
 
-```jldoctest
+```gfdoctest
 using IndividualDisplacements, Statistics
 p=dirname(pathof(IndividualDisplacements))
 include(joinpath(p,"../examples/worldwide/three_dimensional_ocean.jl"))
@@ -161,7 +161,7 @@ end
 Interpolate velocity from gridded fields (2D; NO halos) to position `u`
 (`x,y`) to compute the derivative of position v time  `du_dt`.
 
-```jldoctest
+```gfdoctest
 using IndividualDisplacements, Statistics
 p=dirname(pathof(IndividualDisplacements))
 include(joinpath(p,"../examples/basics/particle_cloud.jl"))
