### 2020-11-02 #2813 @whedon
[REVIEW]: IndividualDisplacements.jl: a Julia package to simulate and study particle displacements within the climate system
**Submitting author:** @gaelforget (<a href="http://orcid.org/0000-0002-4234-056X">Gael Forget</a>)
**Repository:** <a href="https://github.com/JuliaClimate/IndividualDisplacements.jl" target ="_blank">https://github.com/JuliaClimate/IndividualDisplacements.jl</a>
**Version:** v0.2.0
