### 2020-11-02 #2813 @kthyng
@milankl, @visr this is where the actual review takes place. As you go through the checklist and the [JOSS reviewer docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), open issues in the software repository, then link to this review issue for bookkeeping. That way, this issue stays relatively clean but has clear links out to all related review issues and they can be tracked independently in other issues. Let me know if you have any questions and otherwise I'll check in periodically.
