+++ .travis.yml
@@ -2,9 +2,7 @@ language: julia
 os:
   - linux
 julia:
-  - 1.3
   - 1.5
-  - nightly
 notifications:
   email: false
 
