+++ .travis.yml
@@ -17,7 +17,7 @@ after_success:
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.3
+      julia: 1.5
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
