### 2020-11-05 f59e650
try reducing memory footprint
Former-commit-id: 5bb235e09791ac5390b5ea7b30ecf57110f0f14a
