+++ examples/helper_functions.jl
@@ -179,21 +179,25 @@ function OCCA_setup(;backward_in_time::Bool=false)
    γ=GridSpec("PeriodicChannel",MeshArrays.GRID_LL360)
    Γ=GridLoad(γ)
    n=length(Γ["RC"])
+   n=10
+
+   delete!.(Ref(Γ), ["hFacC", "hFacW", "hFacS","DXG","DYG","RAC","RAZ","RAS"]);
+   backward_in_time ? s=-1.0 : s=1.0
 
    fileIn=OCCAclim_path*"DDuvel.0406clim.nc"
-   u = ncread(fileIn,"u")
+   u = ncread(fileIn,"u")[:,:,1:n,:]
    u=dropdims(mean(u,dims=4),dims=4)
    u[findall(u .< -1.0e10)] .=0.0
-   u=read(u,MeshArray(γ,Float32,n))
+   u=s*read(u,MeshArray(γ,Float32,n))
 
    fileIn=OCCAclim_path*"DDvvel.0406clim.nc"
-   v = ncread(fileIn,"v")
+   v = ncread(fileIn,"v")[:,:,1:n,:]
    v=dropdims(mean(v,dims=4),dims=4)
    v[findall(v .< -1.0e10)] .=0.0
    v=read(v,MeshArray(γ,Float32,n))
 
    fileIn=OCCAclim_path*"DDwvel.0406clim.nc"
-   w = ncread(fileIn,"w")
+   w = ncread(fileIn,"w")[:,:,1:n,:]
    w=dropdims(mean(w,dims=4),dims=4)
    w[findall(w .< -1.0e10)] .=0.0
    w=-cat(w,zeros(360, 160),dims=3)
@@ -201,16 +205,16 @@ function OCCA_setup(;backward_in_time::Bool=false)
    w=read(w,MeshArray(γ,Float32,n+1))
 
    fileIn=OCCAclim_path*"DDtheta.0406clim.nc"
-   θ = ncread(fileIn,"theta")
+   θ = ncread(fileIn,"theta")[:,:,1:n,:]
    θ=dropdims(mean(θ,dims=4),dims=4)
    θ[findall(θ .< -1.0e10)] .=NaN
    θ=read(θ,MeshArray(γ,Float32,n))
 
-   fileIn=OCCAclim_path*"DDsalt.0406clim.nc"
-   𝑆 = ncread(fileIn,"salt")
-   𝑆=dropdims(mean(𝑆,dims=4),dims=4)
-   𝑆[findall(𝑆 .< -1.0e10)] .=NaN
-   𝑆=read(𝑆,MeshArray(γ,Float32,n))
+#   fileIn=OCCAclim_path*"DDsalt.0406clim.nc"
+#   𝑆 = ncread(fileIn,"salt")[:,:,1:n,:]
+#   𝑆=dropdims(mean(𝑆,dims=4),dims=4)
+#   𝑆[findall(𝑆 .< -1.0e10)] .=NaN
+#   𝑆=read(𝑆,MeshArray(γ,Float32,n))
 
    for i in eachindex(u)
       u[i]=u[i]./Γ["DXC"][1]
@@ -221,7 +225,7 @@ function OCCA_setup(;backward_in_time::Bool=false)
       u[i]=circshift(u[i],[-180 0])
       v[i]=circshift(v[i],[-180 0])
       θ[i]=circshift(θ[i],[-180 0])
-      𝑆[i]=circshift(𝑆[i],[-180 0])
+#      𝑆[i]=circshift(𝑆[i],[-180 0])
    end
 
[…]
