### 2020-11-05 83a38fc
reduce problem sizes & try early downloads
