### 2020-11-05 302ab7d
reduce problem sizes & try early downloads
Former-commit-id: 83a38fce6f8be92821b8da65e79dcd77b045ed88
