### 2020-11-05 81acde8
reduce problem sizes & try early downloads
