+++ .travis.yml
@@ -8,6 +8,13 @@ notifications:
 
 script: # the default script is equivalent to the following
   - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+
+  - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.build();'
+  - julia --project -e 'using IndividualDisplacements, DataFrames, OceanStateEstimation, NetCDF;
+                        include("examples/helper_functions.jl"));
+                        get_ecco_velocity_if_needed();
+                        get_occa_velocity_if_needed();
+                        get_flt_ex_if_needed();';
   - travis_wait 30 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
 
 after_success:
