### 2020-11-06 453bc86
lazy read via NetCDF.open
