### 2020-11-06 d2f3d25
add (untested) dxyz_dt! and skip ∫ in 3D example
