### 2020-11-06 4c44a10
add (untested) dxyz_dt! and skip ∫ in 3D example
Former-commit-id: d2f3d250bccd56faff5642fceebd28aab8fc34f0
