### 2020-11-06 d254a94
switch set_up_individuals off , but now use dxyz_dt!
Former-commit-id: 103e4c1511e6d402481aab84aaf5c029cdfca33d
