### 2020-11-06 103e4c1
switch set_up_individuals off , but now use dxyz_dt!
