+++ examples/worldwide/three_dimensional_ocean.jl
@@ -33,7 +33,7 @@ bck=false
 
 if nam=="OCCA"
    𝑃,Γ=OCCA_setup(backward_in_time=bck)
-   🚄 =dxyz_dt
+   🚄 =dxyz_dt!
 elseif nam=="LL90"
    𝑃,Γ=example3_setup(backward_in_time=bck)
    🚄 =dxy_dt
@@ -101,6 +101,7 @@ function set_up_individuals(𝑃,Γ,∫,🚄,🔧; nf=10000,
    lat=la0 .+(la1-la0).*rand(nf)
    (xy,_)=initialize_lonlat(Γ,lon,lat)
    xy[3,:] .= z_init
+   xy=cat(xy,ones(1,10),dims=1)
    id=collect(1:size(xy,2))
 
    tr = DataFrame([fill(Int, 2) ; fill(Float64, 9); fill(Symbol, 1)], 
@@ -113,7 +114,7 @@ end
 
 set_up_individuals(𝐼::Individuals; nf=10000) = set_up_individuals(𝑃,Γ,∫,🚄,🔧; nf=nf)
 
-𝐼=set_up_individuals(𝑃,Γ,∫,🚄,🔧,nf=10)
+#𝐼=set_up_individuals(𝑃,Γ,∫,🚄,🔧,nf=10)
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
 # ## 3.1 Compute Displacements
