### 2020-11-06 804903b
replace initialization part
