+++ examples/worldwide/three_dimensional_ocean.jl
@@ -127,7 +127,7 @@ set_up_individuals(𝐼::Individuals; nf=10000) = set_up_individuals(𝑃,Γ,∫
 
 𝑇=(0.0,100*86400.0)
 
-#∫!(𝐼,𝑇)
+∫!(𝐼,𝑇)
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
 # ## 3.2 Analyze Results
