### 2020-11-06 b56ebb1
get grids and velocities via Artifacts, and early on [1/n]
Former-commit-id: 6c2f7ffd00e76e5e0dfeb5afea7ee71c87d60ea6
