+++ docs/make.jl
@@ -6,9 +6,9 @@ module datadeps
   using IndividualDisplacements, OceanStateEstimation
   include(joinpath(dirname(pathof(IndividualDisplacements)),"../examples/helper_functions.jl"))
   include(joinpath(dirname(pathof(IndividualDisplacements)),"../test/helper_functions.jl"))
-  get_llc90_grid_if_needed(); get_ecco_velocity_if_needed();
-  get_ll360_grid_if_needed(); get_occa_velocity_if_needed();
-  get_flt_ex_if_needed();
+  #uget_ecco_velocity_if_needed();
+  #get_occa_velocity_if_needed();
+  #get_flt_ex_if_needed();
 end
 
 # generate tutorials and how-to guides using Literate
@@ -21,8 +21,9 @@ nb = true      # Set to true to generate the notebooks
 
 lst1 = ["solid_body_rotation","random_flow_field","global_ocean_circulation","three_dimensional_ocean","detailed_look","particle_cloud"]
 lst2 = ["solid_body_rotation","random_flow_field","global_ocean_circulation","three_dimensional_ocean","detailed_look","particle_cloud"]
-tst1(x) = Bool(sum(isequal.(x, lst1)))
-tst2(x) = Bool(sum(isequal.(x, lst2)))
+#lst2 = ["none"]
+tst1(x) = !isempty(lst1) && Bool(sum(isequal.(x, lst1)))
+tst2(x) = !isempty(lst2) && Bool(sum(isequal.(x, lst2)))
 
 for (root, _, files) in walkdir(lit), file in files
     splitext(file)[2] == ".jl" || continue
