### 2020-11-06 c455dd2
get grids and velocities via Artifacts, and early on [1/n]
Former-commit-id: b56ebb14f255203db0db77cf01a0aca2c009f35a [formerly 6c2f7ffd00e76e5e0dfeb5afea7ee71c87d60ea6]
Former-commit-id: 978fc3e20d19960088e1817ba34d6e00b8e46bec
Former-commit-id: 7050b2e8428466787cd54925bdc183cbb7e55bd5
