+++ .travis.yml
@@ -2,14 +2,19 @@ language: julia
 os:
   - linux
 julia:
-  - 1.2
   - 1.5
-  - nightly
 notifications:
   email: false
 
 script: # the default script is equivalent to the following
   - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
+
+  - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.build();'
+  - julia --project -e 'using IndividualDisplacements, DataFrames, OceanStateEstimation, NetCDF;
+                        include("examples/helper_functions.jl"));
+                        get_ecco_velocity_if_needed();
+                        get_occa_velocity_if_needed();
+                        get_flt_ex_if_needed();';
   - travis_wait 30 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
 
 after_success:
@@ -19,12 +24,12 @@ after_success:
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.2
+      julia: 1.5
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
                                     Pkg.develop(PackageSpec(path=pwd()))'
-        - julia --project=docs/ docs/make.jl
+        - travis_wait 40 julia --project=docs/ docs/make.jl
       after_success: skip
   allow_failures:
     - julia: nightly
