### 2020-11-06 290a501
comment out plotting except in md mode
Former-commit-id: 7dd9b8208c60928615ba460c80b06c7872dae79a [formerly 14dca4c6a45412936cfab365536fe02bc6f5d797]
Former-commit-id: ba72862a31472835a3ba9ac7c771461ba6e7ba50
Former-commit-id: 44ea25a2a25370e260ea811ea10de76e71873af7
