+++ docs/make.jl
@@ -6,7 +6,7 @@ module datadeps
   using IndividualDisplacements, OceanStateEstimation
   include(joinpath(dirname(pathof(IndividualDisplacements)),"../examples/helper_functions.jl"))
   include(joinpath(dirname(pathof(IndividualDisplacements)),"../test/helper_functions.jl"))
-  #uget_ecco_velocity_if_needed();
+  #get_ecco_velocity_if_needed();
   #get_occa_velocity_if_needed();
   #get_flt_ex_if_needed();
 end
