### 2020-11-06 14dca4c
comment out plotting except in md mode
