### 2020-11-06 b520b16
use Artifacts [2/2], fix slow init & improce read in OCCA_setup
