### 2020-11-06 bfee5d3
use Artifacts [2/2], fix slow init & improce read in OCCA_setup
Former-commit-id: 3e1e0e0f744eb8653d271e825abb8859bad77c9d [formerly b520b16cbd0ae4ac627224ae45d9a30dc5336f4a]
Former-commit-id: f5d533d400e469cca75aa6ef515fdc959168a03b
Former-commit-id: 11161d0ef822aa133fce328ab02f47569f4cacde
