### 2020-11-06 3e1e0e0
use Artifacts [2/2], fix slow init & improce read in OCCA_setup
Former-commit-id: b520b16cbd0ae4ac627224ae45d9a30dc5336f4a
