+++ examples/helper_functions.jl
@@ -1,46 +1,16 @@
 
 """
-    get_llc90_grid_if_needed()
-
-Download global `MITgcm` grid and transport output to `examples/GRID_LLC90`
-"""
-function get_llc90_grid_if_needed()
-  p=dirname(pathof(IndividualDisplacements))
-  p=joinpath(p,"../examples/GRID_LLC90")
-  r="https://github.com/gaelforget/GRID_LLC90"
-  !isdir(p) ? run(`git clone $r $p`) : nothing
-end
-
-"""
     get_ecco_velocity_if_needed()
 
 Download `MITgcm` transport output to `examples/nctiles_climatology` if needed
 """
 function get_ecco_velocity_if_needed()
-    p=dirname(pathof(IndividualDisplacements))
-    q=dirname(pathof(OceanStateEstimation))
-    pth0=pwd()
-    cd(joinpath(p,"../examples/"))
-    lst="nctiles_climatology.csv"
-    pth="nctiles_climatology/"
-    !isfile(lst) ? run(`cp $q/../examples/nctiles_climatology.csv $p/../examples/`) : nothing
-    !isdir("$pth") ? mkdir("$pth") : nothing
-    !isdir("$pth"*"UVELMASS") ? get_from_dataverse(lst,"UVELMASS",pth) : nothing
-    !isdir("$pth"*"VVELMASS") ? get_from_dataverse(lst,"VVELMASS",pth) : nothing
-    !isdir("$pth"*"WVELMASS") ? get_from_dataverse(lst,"WVELMASS",pth) : nothing
-    cd(pth0)
-end
-
-"""
-    get_ll360_grid_if_needed()
-
-Download global `MITgcm` grid and transport output to `examples/GRID_LL360`
-"""
-function get_ll360_grid_if_needed()
-  p=dirname(pathof(IndividualDisplacements))
-  p=joinpath(p,"../examples/GRID_LL360")
-  r="https://github.com/gaelforget/GRID_LL360"
-  !isdir(p) ? run(`git clone $r $p`) : nothing
+    p=dirname(pathof(OceanStateEstimation))
+    lst=joinpath(p,"../examples/nctiles_climatology.csv")
+    pth=ECCOclim_path
+    !isdir(pth*"UVELMASS") ? get_from_dataverse(lst,"UVELMASS",pth) : nothing
+    !isdir(pth*"VVELMASS") ? get_from_dataverse(lst,"VVELMASS",pth) : nothing
+    !isdir(pth*"WVELMASS") ? get_from_dataverse(lst,"WVELMASS",pth) : nothing
 end
 
 """
@@ -49,22 +19,16 @@ end
 Download `MITgcm` transport output to `examples/OCCA_climatology` if needed
 """
 function get_occa_velocity_if_needed()
-    p=dirname(pathof(IndividualDisplacements))
-    q=dirname(pathof(OceanStateEstimation))
-    pth0=pwd()
-    cd(joinpath(p,"../examples/"))
-    lst="OCCA_climatology.csv"
-    pth="OCCA_climatology/"
-    !isfile(lst) ? run(`cp $q/../examples/OCCA_climatology.csv $p/../examples/`) : nothing
[…]
