### 2020-11-06 fba0dc2
skip computation of examples in docs
