### 2020-11-06 ec4051b
tweak CI params
