+++ .travis.yml
@@ -2,6 +2,7 @@ language: julia
 os:
   - linux
 julia:
+  - 1.3
   - 1.5
 notifications:
   email: false
@@ -29,7 +30,7 @@ jobs:
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
                                     Pkg.develop(PackageSpec(path=pwd()))'
-        - travis_wait 40 julia --project=docs/ docs/make.jl
+        - travis_wait 59 julia --project=docs/ docs/make.jl
       after_success: skip
   allow_failures:
     - julia: nightly
