### 2020-11-07 cb1f35a
mv get*if*needed* functions to src
