+++ test/runtests.jl
@@ -2,14 +2,9 @@ using Test, Documenter
 using IndividualDisplacements, MeshArrays, OrdinaryDiffEq, OceanStateEstimation
 include(joinpath(dirname(pathof(IndividualDisplacements)),"../test/helper_functions.jl"))
 
-module datadeps
-  using IndividualDisplacements, OceanStateEstimation
-  include(joinpath(dirname(pathof(IndividualDisplacements)),"../test/helper_functions.jl"))
-  include(joinpath(dirname(pathof(IndividualDisplacements)),"../examples/helper_functions.jl"))
-  get_ecco_velocity_if_needed();
-  get_occa_velocity_if_needed();
-  get_flt_ex_if_needed();
-end;
+IndividualDisplacements.get_ecco_velocity_if_needed();
+IndividualDisplacements.get_occa_velocity_if_needed();
+IndividualDisplacements.get_flt_ex_if_needed();
 
 @testset "test1" begin
     uvetc,sol=test1_setup()
