+++ src/read.jl
@@ -151,3 +151,49 @@ function read_uvetc(k::Int,Γ::Dict,pth::String)
 
     return merge(𝑃,tmp)
 end
+
+
+"""
+    get_ecco_velocity_if_needed()
+
+Download `MITgcm` transport output to `examples/nctiles_climatology` if needed
+"""
+function get_ecco_velocity_if_needed()
+    p=dirname(pathof(OceanStateEstimation))
+    lst=joinpath(p,"../examples/nctiles_climatology.csv")
+    pth=ECCOclim_path
+    !isdir(pth*"UVELMASS") ? get_from_dataverse(lst,"UVELMASS",pth) : nothing
+    !isdir(pth*"VVELMASS") ? get_from_dataverse(lst,"VVELMASS",pth) : nothing
+    !isdir(pth*"WVELMASS") ? get_from_dataverse(lst,"WVELMASS",pth) : nothing
+end
+
+"""
+    get_occa_velocity_if_needed()
+
+Download `MITgcm` transport output to `examples/OCCA_climatology` if needed
+"""
+function get_occa_velocity_if_needed()
+    p=dirname(pathof(OceanStateEstimation))
+    lst=joinpath(p,"../examples/OCCA_climatology.csv")
+    pth=OCCAclim_path
+    nams = ("DDuvel.0406clim.nc","DDvvel.0406clim.nc","DDwvel.0406clim.nc","DDtheta.0406clim.nc","DDsalt.0406clim.nc")
+    if !isfile("$pth"*"DDuvel.0406clim.nc") 
+        tmp=joinpath(pth,"tmp/")
+        !isdir(tmp) ? mkdir(tmp) : nothing
+        [get_from_dataverse(lst,nam,tmp) for nam in nams]
+        [mv(joinpath(tmp,nam,nam),joinpath(pth,nam)) for nam in nams]
+    end
+end
+
+"""
+    get_flt_ex_if_needed()
+
+Download simple grid, velocity, & trajectory output from `MITgcm/pkg/flt`
+to `examples/flt_example`
+"""
+function get_flt_ex_if_needed()
+    p=dirname(pathof(IndividualDisplacements))
+    p=joinpath(p,"../examples/flt_example")
+    r="https://github.com/gaelforget/flt_example"
+    !isdir(p) ? run(`git clone $r $p`) : nothing
+end
