+++ .travis.yml
@@ -11,11 +11,10 @@ script: # the default script is equivalent to the following
   - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
 
   - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.build();'
-  - julia --project -e 'using IndividualDisplacements, DataFrames, OceanStateEstimation, NetCDF;
-                        include("examples/helper_functions.jl"));
-                        get_ecco_velocity_if_needed();
-                        get_occa_velocity_if_needed();
-                        get_flt_ex_if_needed();';
+  - julia --project -e 'using IndividualDisplacements;
+                        IndividualDisplacements.get_ecco_velocity_if_needed();
+                        IndividualDisplacements.get_occa_velocity_if_needed();
+                        IndividualDisplacements.get_flt_ex_if_needed();';
   - travis_wait 30 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
 
 after_success:
