### 2020-11-07 c007b56
mv get*if*needed* functions to src
Former-commit-id: cb1f35a553b2e28107340f906ae9eebf38efa139
