### 2020-11-07 e48c1d5
mv get*if*needed* functions to src
Former-commit-id: c007b56e2df8e5bbb55ca50ff5118700c7160971 [formerly cb1f35a553b2e28107340f906ae9eebf38efa139]
Former-commit-id: fc79b851d892603fd5369ea04817220c95f1a67e
Former-commit-id: 33cc4fbf24cb6ce0a854bc01dd774201b6d60f2e
