### 2020-11-07 aed68e2
add OceanStateEstimation back
Former-commit-id: 64aea0098d944f6751d2d0f7474a965fce99e2e7
