### 2020-11-07 b231724
call get*if*needed* functions preemptively
Former-commit-id: 92950b3f83bf545f686125abd470a1518b9ed473 [formerly f7636494bf24580f6baf6fbfa48ef362d3655942]
Former-commit-id: 7f1c4b143805cc4e53cda8667560ab83cdb7dd30
Former-commit-id: 2eed4a85c958099dd47ac7d1a1a4db154a94b0f2
