+++ examples/basics/detailed_look.jl
@@ -34,7 +34,8 @@ include(joinpath(p,"../examples/helper_functions.jl"));
 IndividualDisplacements.get_flt_ex_if_needed()
 dirIn=joinpath(p,"../examples/flt_example/")
 prec=Float32
-df=read_flt(dirIn,prec)
+df=read_flt(dirIn,prec);
+
 #!md plt=PlotBasic(df,300,100000.0)
 
 # ## 3. Read Gridded Variables
@@ -60,7 +61,8 @@ tmp[1:4,:]
 
 x=Γ["XG"].f[1][:,1]
 y=Γ["YC"].f[1][1,:]
-z=transpose(Γ["mskW"][1].*𝑃.u0[1,1])
+z=transpose(Γ["mskW"][1].*𝑃.u0[1,1]);
+
 #!md plt=contourf(x,y,z,c=:delta)
 #!md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
 
@@ -68,7 +70,8 @@ z=transpose(Γ["mskW"][1].*𝑃.u0[1,1])
 
 x=Γ["XC"].f[1][:,1]
 y=Γ["YG"].f[1][1,:]
-z=transpose(Γ["mskW"][1].*𝑃.v0[1,1])
+z=transpose(Γ["mskW"][1].*𝑃.v0[1,1]);
+
 #!md plt=contourf(x,y,z,c=:delta)
 #!md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
 
@@ -89,6 +92,7 @@ for i=1:100
     tmpu[i]=du[1]
     tmpv[i]=du[2]
 end
+
 #!md plt=plot(tmpx,tmpu,label="u (interp)")
 #!md plot!(Γ["XG"].f[1][1:10,1]./𝑃.dx,𝑃.u0.f[1][1:10,1],marker=:o,label="u (C-grid)")
 #!md plot!(tmpx,tmpv,label="v (interp)")
@@ -105,6 +109,7 @@ for i=1:100
     tmpu[i]=du[1]
     tmpv[i]=du[2]
 end
+
 #!md plt=plot(tmpx,tmpu,label="u (interp)")
 #!md plot!(Γ["YG"].f[1][1,1:10]./𝑃.dx,𝑃.u0.f[1][1,1:10],marker=:o,label="u (C-grid)")
 #!md plot!(tmpx,tmpv,label="v (interp)")
@@ -124,7 +129,7 @@ for i=1:nSteps
     tmpu[i]=du[1]
     tmpv[i]=du[2]
 end
-#
+
 #!md plt=plot(tmpu,label="u")
 #!md plot!(tmpv,label="v")
 #!md plot!(refu,label="u (ref)")
@@ -159,7 +164,7 @@ for i=1:nSteps-1
     ref[2,i+1]-ref[2,i]>maxLat/2 ? ref[2,i+1:end]-=fill(maxLat,(nSteps-i)) : nothing
     ref[2,i+1]-ref[2,i]<-maxLat/2 ? ref[2,i+1:end]+=fill(maxLat,(nSteps-i)) : nothing
 end
-ref=ref./𝑃.dx
+ref=ref./𝑃.dx;
 
 #!md plt=plot(sol[1,:],sol[2,:],linewidth=5,title="Using Recomputed Velocities",
 #!md      xaxis="lon",yaxis="lat",label="Julia Solution") # legend=false
