### 2020-11-07 04ca17e
bump version and fix docs build (?)
