### 2020-11-07 b72577e
bump version and fix docs build (?)
Former-commit-id: ec88b0c8daeaa34aa6c4aaf1866be1d9d5928dd5 [formerly 04ca17ea95dbde9875858d68996d24886828e609]
Former-commit-id: b40ad17c43d9911070610122730a83ba3e6de4ae
Former-commit-id: da556142c63c056de472e376f14d9b323873a45e
