+++ examples/example123.jl
@@ -61,8 +61,7 @@ example2_xy()
 Read MITgcm/pkg/flt output
 """
 function example2_xy(𝑃)
-p=dirname(pathof(IndividualDisplacements))
-dirIn=joinpath(p,"../examples/flt_example/")
+dirIn=IndividualDisplacements.flt_example
 prec=Float32
 df=read_flt(dirIn,prec)
 #
@@ -92,8 +91,7 @@ function example2_setup()
 
    ###### 1) Get gridded variables via MeshArrays.jl
 
-   p=dirname(pathof(IndividualDisplacements))
-   dirIn=joinpath(p,"../examples/flt_example/")
+   dirIn=IndividualDisplacements.flt_example
    γ=gcmgrid(dirIn,"PeriodicChannel",1,[(80,42)], [80 42], Float32, read, write)
    nr=8
 
