+++ .travis.yml
@@ -13,8 +13,7 @@ script: # the default script is equivalent to the following
   - julia --project -e 'using Pkg; Pkg.instantiate(); Pkg.build();'
   - julia --project -e 'using IndividualDisplacements;
                         IndividualDisplacements.get_ecco_velocity_if_needed();
-                        IndividualDisplacements.get_occa_velocity_if_needed();
-                        IndividualDisplacements.get_flt_ex_if_needed();';
+                        IndividualDisplacements.get_occa_velocity_if_needed();';
   - travis_wait 30 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
 
 after_success:
@@ -31,8 +30,7 @@ jobs:
                                     Pkg.develop(PackageSpec(path=pwd()))'
         - julia --project=docs/ -e 'using IndividualDisplacements;
                                     IndividualDisplacements.get_ecco_velocity_if_needed();
-                                    IndividualDisplacements.get_occa_velocity_if_needed();
-                                    IndividualDisplacements.get_flt_ex_if_needed();';
+                                    IndividualDisplacements.get_occa_velocity_if_needed();';
         - travis_wait 59 julia --project=docs/ docs/make.jl
       after_success: skip
   allow_failures:
