### 2020-11-07 c108c50
add detailed_look, particle_cloud back
Former-commit-id: db941dcb42662ee885c49767f0061fa5f429adbc
