### 2020-11-07 f97a418
add nb computation back
