### 2020-11-07 692d854
add in place 3D code
Former-commit-id: 27007beff476b0ccc66ba67fd62d3e29805ba797
