+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"EditURL = \"https://github.com/JuliaClimate/IndividualDisplacements.jl/blob/master/examples/worldwide/global_ocean_circulation.jl\"","category":"page"},{"location":"worldwide/global_ocean_circulation/#Global-Climatology","page":"Global Climatology","title":"Global Climatology","text":"","category":"section"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"(Image: ) (Image: )","category":"page"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"Advect particles with climatological monthly mean flow at selected depth level (e.g. k=10 for 95 m) from a global ocean state estimate (ECCO v4 r2 ; see also https://ecco-group.org) which is here repeated for ny years. For additional documentation e.g. see : 1, 2, 3, 4","category":"page"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"(Image: simulated particle movie (5m))","category":"page"},{"location":"worldwide/global_ocean_circulation/#.-Get-Software-and-Iput-Files","page":"Global Climatology","title":"1. Get Software & Iput Files","text":"","category":"section"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"packages + helper functions\ngrid and velocity files","category":"page"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"using IndividualDisplacements, MeshArrays, OrdinaryDiffEq\nusing Statistics, DataFrames, MITgcmTools, OceanStateEstimation\n\ninclude(joinpath(dirname(pathof(IndividualDisplacements)),\"../examples/helper_functions.jl\"))\nget_llc90_grid_if_needed(); get_ecco_velocity_if_needed();\nnothing #hide","category":"page"},{"location":"worldwide/global_ocean_circulation/#.-Set-Up-Parameters-and-Inputs","page":"Global Climatology","title":"2. Set Up Parameters & Inputs","text":"","category":"section"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"select vertical level & duration in years\nread grid variables & velocities\nnormalize velocities","category":"page"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"𝑃=setup_global_ocean(k=1,ny=2);\n\nkeys(𝑃)","category":"page"},{"location":"worldwide/global_ocean_circulation/#.-Main-Computation-Loop","page":"Global Climatology","title":"3. Main Computation Loop","text":"","category":"section"},{"location":"worldwide/global_ocean_circulation/#.1-Initialize-Individuals-and-Solution","page":"Global Climatology","title":"3.1 Initialize Individuals & Solution","text":"","category":"section"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"initial particle positions randomly over Global Ocean","category":"page"},{"location":"worldwide/global_ocean_circulation/","page":"Global Climatology","title":"Global Climatology","text":"xy=init
[…]
