### 2020-11-07 532fe8a
Update API.md
Former-commit-id: 98d0f32314fbca6852ce2543511c2f99eb4a5f6e [formerly 55cac2274ee91786545f91f1008056b7acd55c81]
Former-commit-id: 8da90f23fd6132e13c230643fd290b9169607871
Former-commit-id: 830c621e7f7b930a9e2ddda4ae27749914741d30
