+++ docs/src/API.md
@@ -1,11 +1,12 @@
 ## Velocity Interpolation
 
-The `dxy_dt`, `dxy_dt!`, `dxyz_dt`, etc functions compute the tracked individual velocity. 
+The `dxy_dt`, `dxy_dt!`, `dxyz_dt`, `dxyz_dt!`, etc functions compute the tracked individual velocity. 
 
 ```@docs
 dxy_dt
 dxy_dt!
 dxyz_dt
+dxyz_dt!
 dxy_dt_replay
 dxy_dt_CyclicArray
 ```
