### 2020-11-16 6f72217
rm un-needed dependency to StatsBase
