### 2020-11-16 476ef12
rm un-needed dependency to StatsBase
Former-commit-id: 27d4aeabc8ee3640bcf65a4f728e0cdad751b0f9 [formerly 6f72217f5656b02f226ead380e7392d94a7908cd]
Former-commit-id: 1f0494c70e671f7ee5e60ce65a185ca0aeb667d1
Former-commit-id: 7e2d8e8d9bd62452bdba4004e508a9e6cb80a244
