+++ Project.toml
@@ -16,7 +16,6 @@ OceanStateEstimation = "891f6deb-a4f5-4bc5-a2e3-1e8f649cdd2c"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
@@ -29,6 +28,5 @@ MeshArrays = "0.2.14"
 NetCDF = "0.10"
 OceanStateEstimation = "0.1.5"
 OrdinaryDiffEq = "5"
-StatsBase = "0.33"
 UnPack = "1.0"
 julia = "^1.3"
