### 2020-11-16 02cac05
add windows and osx CI
Former-commit-id: ba6230afa08111ae68d8279925f5600bdf41b3a4 [formerly 9fbf0c431d6876839e5491ae7d0e408294356c57]
Former-commit-id: 539ddea1211e167f12bbe7b27a54c33bb8b57b5c
Former-commit-id: e775eec3f183f144a194e8a4fd5e320a02ce281e
