+++ .travis.yml
@@ -1,6 +1,8 @@
 language: julia
 os:
   - linux
+  - osx
+  - windows
 julia:
   - 1.3
   - 1.5
