### 2020-11-16 3e3bbcd
comment out windows CI (Plots doesnt build)
Former-commit-id: 0e2d64b21171662ea948ed43dd872495fdb689d4 [formerly 64d9b2f797fade7c6f263363ceab0558c535ac5a]
Former-commit-id: 62e9357ce8bcdd4145437ee2f99b388edddb3152
Former-commit-id: 5e4888471f684109e812d713e5c2d104343082f7
