+++ .travis.yml
@@ -2,7 +2,6 @@ language: julia
 os:
   - linux
   - osx
-  - windows
 julia:
   - 1.3
   - 1.5
