### 2020-11-17 36717b0
set travis wait to 40
Former-commit-id: e6deafc238492147fe733f7a2f7b61196c680306 [formerly 53bbcaf0dccc7c6591dd9a34043efabb61a6e280]
Former-commit-id: 799dfca102556b303b34c806b70282ed82ca4f5a
Former-commit-id: 809286ccdc61a1466d33c86f1e62d02d2e56de78
