+++ .travis.yml
@@ -15,7 +15,7 @@ script: # the default script is equivalent to the following
   - julia --project -e 'using IndividualDisplacements;
                         IndividualDisplacements.get_ecco_velocity_if_needed();
                         IndividualDisplacements.get_occa_velocity_if_needed();';
-  - travis_wait 30 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
+  - travis_wait 40 julia --project -e 'using Pkg; Pkg.build(); Pkg.test(; coverage=true)';
 
 after_success:
   - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())';
@@ -32,7 +32,7 @@ jobs:
         - julia --project=docs/ -e 'using IndividualDisplacements;
                                     IndividualDisplacements.get_ecco_velocity_if_needed();
                                     IndividualDisplacements.get_occa_velocity_if_needed();';
-        - travis_wait 59 julia --project=docs/ docs/make.jl
+        - travis_wait 40 julia --project=docs/ docs/make.jl
       after_success: skip
   allow_failures:
     - julia: nightly
