### 2020-11-17 a39aec9
skip osx travis (expires on 1.5 ...)
Former-commit-id: af403b9193248998dd6ed389809dd11b7ef21bf1 [formerly 7a5df040c6674013deb461f5fe6bd3feb21d3ba4]
Former-commit-id: 4e86e0a20e59b31923d870f9ccf993d6ed7f98b8
Former-commit-id: 2f3c243af9f95ac91aeafb335d3e17ca07e7b66d
