+++ .travis.yml
@@ -1,7 +1,6 @@
 language: julia
 os:
   - linux
-  - osx
 julia:
   - 1.3
   - 1.5
