+++ Project.toml
@@ -22,7 +22,7 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 CFTime = "0.1"
 CSV = "0.6, 0.7, 0.8"
 CyclicArrays = "0.2, 0.3, 0.4"
-DataFrames = "0.21, 0.22"
+DataFrames = "0.21"
 MITgcmTools = "0.1.6"
 MeshArrays = "0.2.14"
 NetCDF = "0.10"
