### 2020-11-17 JuliaClimate/IndividualDisplacements.jl#38 @milankl
Documentation: How to provide C-grid u,v?
Hey Gaël, going through the documentation, I find in most examples that you've hidden the explanation for how to set up the NamedTuple Parameter 𝑃 which contains field like `u0`,`u1`,`v0`,`v1` etc. in some underlying function like `𝑃,Γ=example2_setup()`. I'd assume that most users will have their velocity fields in some form, e.g. simply as julia array loaded from netcdf, so I'd be great to guide them how they can pass these arrays (which may be defined on a C-grid etc.) into IndividualDisplacements.

Yes, you mention MeshArrays.jl, but unfortunately reading that documentation I still don't know how I can, for example, take data from ShallowWaters.jl and compute the particle trajectories. My idea was to test IndividualDisplacements against the semi-Lagrangian tracer advection scheme therein.
