### 2020-11-17 JuliaClimate/IndividualDisplacements.jl#38 @milankl
In the example for Single Particle you write `𝑃=(u0=uu, u1=uu, v0=vv, v1=vv,w0=0.0*w, w1=-0.01*w, 𝑇=[0,19.95*2*pi], ioSize=(np,np,nz))` but that's not entirely helpful either.
