### 2020-11-18 JuliaClimate/IndividualDisplacements.jl#38 @milankl
Maybe as a suggestion would it be possible to add something like
