### 2020-11-24 a705107
rm type restriction, update constructor calls
Former-commit-id: c9fbc393bc4e37cb6ec5593a35e589fbb3802680 [formerly 8d79b77fee0561495081eae73eb40228f5e57be7]
Former-commit-id: 49e3db4279ca60d2d266146e5842499d5c9a356d
Former-commit-id: 69d358028d565f1b3ce8f8cd6193e0482d15e963
