### 2020-11-24 c9fbc39
rm type restriction, update constructor calls
Former-commit-id: 8d79b77fee0561495081eae73eb40228f5e57be7
