### 2020-11-24 ccfad28
reformat position array -- 2D array (2 x n) -> array (1 x n) of 1D array (1 x 2) which impacts initialization (see examples), postprocessing and reduce allocations
Former-commit-id: c04d69d0560175ddb65bc6c875e497562f36dd87 [formerly 7c8ef5a633b70efdca9e60e46cac408fda97c1e3]
Former-commit-id: cacb8f5f1472e58d1bcca9fe1b64b5056db71b57
Former-commit-id: 3c92b477ca10697eaa0e6a0fe95be1f005d090fc
