### 2020-11-24 7c8ef5a
reformat position array -- 2D array (2 x n) -> array (1 x n) of 1D array (1 x 2) which impacts initialization (see examples), postprocessing and reduce allocations
