### 2020-11-24 c04d69d
reformat position array -- 2D array (2 x n) -> array (1 x n) of 1D array (1 x 2) which impacts initialization (see examples), postprocessing and reduce allocations
Former-commit-id: 7c8ef5a633b70efdca9e60e46cac408fda97c1e3
