+++ examples/basics/random_flow_field.jl
@@ -46,7 +46,7 @@ n=100
 x0=x0 .+(x1-x0).*rand(n)
 y0=y0 .+(y1-y0).*rand(n)
 
-xy=transpose([x0[:] y0[:] ones(size(x0[:]))]);
+xy = permutedims([[x0[i];y0[i];1.0] for i in eachindex(x0)])
 
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ## 2.1 Compute Trajectories
@@ -54,7 +54,7 @@ xy=transpose([x0[:] y0[:] ones(size(x0[:]))]);
 tr = DataFrame(ID=Int[], x=Float64[], y=Float64[], t=Float64[])
 solv(prob) = solve(prob,Tsit5(),reltol=1e-5,abstol=1e-5)
 
-I=(position=xy,record=deepcopy(tr),velocity=dxy_dt!,
+I=(position=deepcopy(xy),record=deepcopy(tr),velocity=dxy_dt!,
    integration=solv,postprocessing=postprocess_xy,parameters=𝑃)
 𝐼=Individuals(I)
                       
