### 2020-11-24 30918e2
fix and move reset_lonlat
Former-commit-id: b652b3a51e21ffac004a99575d31ebb13c518cc3
