### 2020-11-24 b652b3a
fix and move reset_lonlat
