+++ examples/worldwide/three_dimensional_ocean.jl
@@ -47,15 +47,26 @@ end
 # ## 2.2 Solver And Analysis Setup
 #
 
-function ∫(prob)
-   sol=solve(prob,Tsit5(),saveat=10*86400.0)
-   nx,ny=𝑃.ioSize[1:2]
-   sol[1,:,:]=mod.(sol[1,:,:],nx)
-   sol[2,:,:]=mod.(sol[2,:,:],ny)
-   return sol
+function my🚄(du::Array{T,2},u::Array{T,2},𝑃::NamedTuple,tim) where T
+   nf=size(u,2)
+   nx=360
+   ny=160
+   [u[1,i][1]=mod(u[1,i][1],nx) for i in 1:nf]
+   [u[1,i][2]=mod(u[1,i][2],ny) for i in 1:nf]
+   [dxyz_dt!(du[i],u[i],𝑃,tim) for i=1:size(u,2)]
 end
 
-∫(prob)=solve(prob,Euler(),dt=86400.0)
+function ∫(prob)
+   #sol=solve(prob,Tsit5(),saveat=10*86400.0)
+   sol=IndividualDisplacements.solver_default(prob)
+   #sol=solve(prob,Euler(),dt=86400.0)
+   nx,ny=𝑃.ioSize[1:2]
+   nf=size(sol,2)
+   nt=size(sol,3)
+   [sol[1,i,j][1]=mod(sol[1,i,j][1],nx) for i in 1:nf, j in 1:nt]
+   [sol[1,i,j][2]=mod(sol[1,i,j][2],ny) for i in 1:nf, j in 1:nt]
+   return sol
+end
 
 function 🔧(sol,𝑃::NamedTuple;id=missing,𝑇=missing)
    df=postprocess_lonlat(sol,𝑃,id=id,𝑇=𝑇)
@@ -116,7 +127,7 @@ function set_up_individuals(𝑃,Γ,∫,🚄,🔧; nf=10000,
                   k=Float64[], z=Float64[], iso=Float64[], t=Float64[], 
                   lon=Float64[], lat=Float64[], year=Float64[], col=Symbol[])
 
-   I=(position=xy,record=deepcopy(tr),velocity=🚄, integration=∫, 
+   I=(position=xy,record=deepcopy(tr),velocity=my🚄, integration=∫, 
       postprocessing=🔧,parameters=𝑃)
    𝐼=Individuals(I)
 
