### 2020-11-25 64c39d3
reorganize and expand docs
Former-commit-id: 5353d8000e08a66195f8515a8f6b29b95e0f1028 [formerly 51c7471ec1ea98c4c7b52346d9515d7a1e322190]
Former-commit-id: cd49fbdfebe69cf1fd1f93c3eeb4197a62e9cefc
Former-commit-id: 19e50969be4eaf25df741b19ee937ebaef1d43f7
