### 2020-11-25 JuliaClimate/IndividualDisplacements.jl#38 @gaelforget
> @gaelforget Could you provide some input to this issue, such that I can continue with the review?

Hi @milankl

Sorry for delay. Have been working on it. More later today on this ...
