### 2020-11-26 dbb1704
improve example guide and solid_body_rotation v docs
Former-commit-id: a39846e4b4cdbf380587f839b8a703a49313f9f9
