### 2020-11-26 42e4aa2
improve example guide and solid_body_rotation v docs
Former-commit-id: dbb170468cd0cfd57dc55c92b50fa99c563c27a5 [formerly a39846e4b4cdbf380587f839b8a703a49313f9f9]
Former-commit-id: af1710200ad5a8b408ac52a0725e2975a5983039
Former-commit-id: d9a4aa5babced3605535ff7fd3830a7272ee5a52
