### 2020-11-26 ec6f74a
revise titles, and add 3d example back but not run
