### 2020-11-26 b648a3f
revise titles, and add 3d example back but not run
Former-commit-id: ec58ffb3616f82fc499a5aa53afc237ab5e54c3f [formerly ec6f74a07a884b9917d55f95ce310dd7d0c149b0]
Former-commit-id: 3511375f4b07c38cada252d56b2b8d556dbd64e3
Former-commit-id: 95f94d305a72d9a62f3e76252b7876d71a45b6d5
