### 2020-11-26 272ef09
documentation fixes and improvements
Former-commit-id: 500d436364b62b26f4c587e0a2e0fe414673688e [formerly 446df4491fb60cc532010c65c6e6357b10440d55]
Former-commit-id: f3e176196f5ec84096f6ca072a368e0018387559
Former-commit-id: 7000c0dc5d40a98c14ed81ee05d62f9faf274cb6
