### 2020-11-26 500d436
documentation fixes and improvements
Former-commit-id: 446df4491fb60cc532010c65c6e6357b10440d55
