### 2020-11-26 82458b0
documentation fixes and improvements 2/n
Former-commit-id: 7e1fad0370a4f0be53b75dd8bc04db1ba93f43a1 [formerly 9cb9b6031cf11f7e49d5e0f8834d135b7540cf04]
Former-commit-id: 9d851cf2154445f7758b9f63fc194140ccd471e5
Former-commit-id: 7a410250f8c4133582608a491f7649fdf256d104
