### 2020-11-26 7e1fad0
documentation fixes and improvements 2/n
Former-commit-id: 9cb9b6031cf11f7e49d5e0f8834d135b7540cf04
