### 2020-11-26 98c157c
documentation fixes and improvements [3/n]
Former-commit-id: f675ddf8667a921de0f4334024208ad36f0f87e0
