### 2020-11-26 f675ddf
documentation fixes and improvements [3/n]
