### 2020-11-27 9100bc8
documentation fixes and improvements [4/n]
Former-commit-id: f4d31bf8fefc8d01be4938c423d8d04f6c6e58a0 [formerly 238b34f2ca47e34fdb111bf2d24e0320c7a77d1a]
Former-commit-id: 70fd99a8f0c6475a28f0665e705645ff9efe07cf
Former-commit-id: 741d2e586170c4b7be8cb27f5fc728c82fbbdbdf
