+++ examples/recipes_plots.jl
@@ -95,18 +95,18 @@ function scatter_movie(𝐼; cam=(0, 90))
 end
 
 """
-    phi_scatter(Γ,ϕ,df)
+    phi_scatter(ϕ,df)
 
 ```
-phi_scatter(Γ,ϕ,df)
+phi_scatter(ϕ,df)
 ```
 """
-function phi_scatter(Γ,ϕ,df)
-    nx,ny=size(ϕ[1])
-    contourf(vec(Γ["XC"][1][:,1]),vec(Γ["YC"][1][1,:]),
-        transpose(ϕ[1]),c = :blues,linewidth = 0.1)
-    scatter!(df.x,df.y,markersize=2.0,c=:red,
-    xlims=(0,nx),ylims=(0,ny),leg=:none,marker = (:circle, stroke(0)))
+function phi_scatter(ϕ,df)
+    nx,ny=size(ϕ)
+    contourf(-0.5 .+ (1:nx),-0.5 .+ (1:ny),
+             transpose(ϕ),c = :blues,linewidth = 0.1)
+    scatter!(df.x,df.y,markersize=2.0,c=:red,marker = (:circle, stroke(0)),
+             xlims=(0,nx),ylims=(0,ny),leg=:none)
 end
 
 """
