### 2020-11-30 JuliaClimate/IndividualDisplacements.jl#38 @gaelforget
Hi @milankl

I have now added documentation to this effect [here](https://juliaclimate.github.io/IndividualDisplacements.jl/dev/examples/) and [here](https://juliaclimate.github.io/IndividualDisplacements.jl/dev/basics/random_flow_field/)

The approach you envisioned (generalizing MeshArray to support non-C grids) is certainly worth considering but would be more involved, so for now I opted for simply adding documentation here.

Does this suffice to allow application to ShallowWaters.jl ?

ps. I hope to spend more time on the docs this week, including the aforementioned sections
