### 2020-11-30 fd237c7
bump version and tweak literate tokens
