### 2020-11-30 f118b31
more documentation improvements
Former-commit-id: 4fa59e83086092c1efb3f350f0a64b644c7b84e3 [formerly e485b1910b682ebccf643b7c09c3a8aadb26c3bc]
Former-commit-id: 0d7675f7560ef27abec24fbc4b8f7a897303d91c
Former-commit-id: 573cc8e354b0310814e9786d965966617fb56a8a
