### 2020-12-01 61084f7
update compat for netcdf
Former-commit-id: 057c0f1f16f958ee375b17c9aa50952bbdf0e59d [formerly 99868955b9c576e8c3eac28ee0ff62fd646d4bea]
Former-commit-id: 9cd39f94bb10f20b2b6ab7e52d352dacdd719607
Former-commit-id: 94cd5573ae27c8fe8737dfe05b0d52b6a2d0d43d
