+++ Project.toml
@@ -25,7 +25,7 @@ CyclicArrays = "0.2, 0.3, 0.4"
 DataFrames = "0.21, 0.22"
 MITgcmTools = "0.1.6"
 MeshArrays = "0.2.14"
-NetCDF = "0.10"
+NetCDF = "0.10, 0.11"
 OceanStateEstimation = "0.1.5"
 OrdinaryDiffEq = "5"
 UnPack = "1.0"
