### 2020-12-01 c101ced
while location_is_out update_location, use Tsit5 by default, etc
Former-commit-id: cf70e512d85bd12b08e801a0a7e9fca666ae2bd4
