+++ examples/basics/detailed_look.jl
@@ -35,7 +35,7 @@ dirIn=IndividualDisplacements.flt_example
 prec=Float32
 df=read_flt(dirIn,prec);
 
-#!md plt=PlotBasic(df,300,100000.0)
+#md plt=PlotBasic(df,300,100000.0)
 
 # ## 3. Read Gridded Variables
 #
@@ -45,9 +45,9 @@ df=read_flt(dirIn,prec);
 
 # ## 4. Visualize Velocity Fields
 
-#!md plt=heatmap(Γ["mskW"][1,1].*𝑃.u0[1,1],title="U at the start")
+#md plt=heatmap(Γ["mskW"][1,1].*𝑃.u0[1,1],title="U at the start")
 
-#!md plt=heatmap(Γ["mskW"][1,1].*𝑃.u1[1,1]-𝑃.u0[1,1],title="U end - U start")
+#md plt=heatmap(Γ["mskW"][1,1].*𝑃.u1[1,1]-𝑃.u0[1,1],title="U end - U start")
 
 # ## 5. Visualize Trajectories
 #
@@ -62,8 +62,8 @@ x=Γ["XG"].f[1][:,1]
 y=Γ["YC"].f[1][1,:]
 z=transpose(Γ["mskW"][1].*𝑃.u0[1,1]);
 
-#!md plt=contourf(x,y,z,c=:delta)
-#!md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
+#md plt=contourf(x,y,z,c=:delta)
+#md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
 
 # Super-impose trajectory over velocity field (... then for v)
 
@@ -71,8 +71,8 @@ x=Γ["XC"].f[1][:,1]
 y=Γ["YG"].f[1][1,:]
 z=transpose(Γ["mskW"][1].*𝑃.v0[1,1]);
 
-#!md plt=contourf(x,y,z,c=:delta)
-#!md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
+#md plt=contourf(x,y,z,c=:delta)
+#md plot!(tmp[:,:lon],tmp[:,:lat],c=:red,w=4,leg=false)
 
 # ## 6. Interpolate Velocities
 
@@ -92,10 +92,10 @@ for i=1:100
     tmpv[i]=du[2]
 end
 
-#!md plt=plot(tmpx,tmpu,label="u (interp)")
-#!md plot!(Γ["XG"].f[1][1:10,1]./𝑃.dx,𝑃.u0.f[1][1:10,1],marker=:o,label="u (C-grid)")
-#!md plot!(tmpx,tmpv,label="v (interp)")
-#!md plot!(Γ["XG"].f[1][1:10,1]./𝑃.dx,𝑃.v0.f[1][1:10,1],marker=:o,label="v (C-grid)")
+#md plt=plot(tmpx,tmpu,label="u (interp)")
+#md plot!(Γ["XG"].f[1][1:10,1]./𝑃.dx,𝑃.u0.f[1][1:10,1],marker=:o,label="u (C-grid)")
+#md plot!(tmpx,tmpv,label="v (interp)")
+#md plot!(Γ["XG"].f[1][1:10,1]./𝑃.dx,𝑃.v0.f[1][1:10,1],marker=:o,label="v (C-grid)")
 
 # And similarly in the other direction
 
@@ -109,10 +109,10 @@ for i=1:100
     tmpv[i]=du[2]
 end
 
-#!md plt=plot(tmpx,tmpu,label="u (interp)")
-#!md plot!(Γ["YG"].f[1][1,1:10]./𝑃.dx,𝑃.u0.f[1][1,1:10],marker=:o,label="u (C-grid)")
[…]
