### 2020-12-01 0bbee10
plots and extend random flow example
