### 2020-12-02 7d5e2ca
set duration to 10 in random flow example
