### 2020-12-02 e76db77
improve doc presentation
Former-commit-id: 8e10adf63db82f9a2be5e8be4bb685a9d446e8fe [formerly 8b1fd945adb4e7859bc57e6425117cb2aa7140e8]
Former-commit-id: f136c64eacbd0cbf28988f99a97736d601949e74
Former-commit-id: 441c9a3a1035faa156199072683df1e363e9bc40
