### 2020-12-04 #2813 @visr
I plan to work on this again next week. Quite a few improvements have come along since I had a look the first time.
