### 2020-12-15 JuliaClimate/IndividualDisplacements.jl#38 @milankl
Then I believe that some functionality as I already mentioned
> ```julia
> uv = MeshArray(u,v,x,y,grid-type=:C,geometry=:rectangle)
> I = Individuals(📌,🔴,uv)
> ```
would greatly increase the usability of this package. I guess ideally a user would want to have functions for three steps
- Provide velocity fields to IndividualDisplacements such they can be correctly interpreted
- Provide particle initial conditions
- Calculate trajectories
I see that you have the `∫!` and the `Individuals` functions for the last two steps, but then I don't understand why in the particle set tutorial you still pass the initial coordinates through the `setup_point_cloud` function?
