### 2020-12-16 0ddbd57
fix 3D case
Former-commit-id: c64801a130185d970fbe3ff8e948d25729de9c58
