### 2020-12-17 f5c3fa4
similar(rec_default) and exp duration
Former-commit-id: 584cf47aa36b5c38e84cfe302e26c4ded1baddfa [formerly c82ddb263a74040f4994cd8d4eaaa385875a4490]
Former-commit-id: f17ecae91c7fda8aafd9e72c0ccd35446141b813
Former-commit-id: 3173bf7c7ec9d370f1881748136954910887c69a
