### 2020-12-17 9442063
turn dxy_dt,dxyz_dt+tests into plain Array option
Former-commit-id: d0a20bfbd70d9133fad345632958678d29ed5196
