### 2020-12-17 0e11723
turn dxy_dt,dxyz_dt+tests into plain Array option
Former-commit-id: 944206360e87b34fc5581cf409e6e75eafb6598a [formerly d0a20bfbd70d9133fad345632958678d29ed5196]
Former-commit-id: d23cdcde49705495ade1adf8dd886a72ec35a7c8
Former-commit-id: 9e7ba8e23e6ad9e7c28845ce1c09b1fd185d84ea
