### 2020-12-17 d0a20bf
turn dxy_dt,dxyz_dt+tests into plain Array option
