### 2020-12-17 3e3b582
new FlowParameters struct [1/n]
Former-commit-id: bad97743826592e14751eed8a0b7b3679469d762 [formerly 66e4e6e067ca40aa9a743e3758bf1a85fb073c8a]
Former-commit-id: 22072490837d4af2ea41d75fc38ff101eafb4840
Former-commit-id: d3d811057583e3e2e957a89c0e63e734ca1ca917
