+++ src/IndividualDisplacements.jl
@@ -16,6 +16,7 @@ include("read.jl")
 include("update_locations.jl")
 
 export Individuals, ∫!, set_up_𝑃, update_𝑃!
+export FlowParameters, 𝑃_Array3D, 𝑃_Array2D, 𝑃_MeshArray3D, 𝑃_MeshArray2D
 export dxy_dt!, dxy_dt, dxyz_dt!, dxyz_dt, dxy_dt_CyclicArray, dxy_dt_replay
 export postprocess_lonlat, add_lonlat!, postprocess_xy, interp_to_xy
 export initialize_gridded, initialize_lonlat, randn_lonlat, interp_to_lonlat
