### 2020-12-17 bad9774
new FlowParameters struct [1/n]
Former-commit-id: 66e4e6e067ca40aa9a743e3758bf1a85fb073c8a
