### 2020-12-17 66e4e6e
new FlowParameters struct [1/n]
