### 2020-12-17 ff6299e
new FlowParameters struct [2/n]
Former-commit-id: 31da6b6f289cfa69bd120daa556e11ae6e7712f4 [formerly 7c1ab7b687654f54048ff5d129a16abea532ebdd]
Former-commit-id: 03b3863ac8f47bea20ca34e61707d7d4733984f8
Former-commit-id: 4d7458aa4b425e9b820155313c61ef9e9251fc08
