### 2020-12-17 31da6b6
new FlowParameters struct [2/n]
Former-commit-id: 7c1ab7b687654f54048ff5d129a16abea532ebdd
