### 2020-12-17 9922a4d
new FlowParameters struct [3/n]
