### 2020-12-17 eb05676
new FlowParameters struct [3/n]
Former-commit-id: 4134e12ff5b8e47183e72b1021666ccfa771e676 [formerly 9922a4d34c27082cbcdf75c3cb255c72a88b5414]
Former-commit-id: fc76aa29caa78e02eda9bb5cc3e5a9edd241530f
Former-commit-id: 3ef1315e6b89d3d980f9272f26634f3441a13b8c
