### 2020-12-18 fbf43a8
new FlowParameters struct [4/n]
Former-commit-id: 7bea74f85ec9299e59bcbe6abeba6d7fe9febf72 [formerly 35a194a8045319f8a1fc6de500e902c9a1586885]
Former-commit-id: aeda00d2371b05283f9d36bd41d63baaafc2f294
Former-commit-id: 66b90ddd52658fc4f207d91a0f81d9385ddf3710
