+++ examples/example123.jl
@@ -132,6 +132,6 @@ function example2_setup()
    mskS=1.0 .+ 0.0 * mask(mskS[:,kk],NaN,0.0)
    Γ=merge(Γ,Dict("mskW" => mskW, "mskS" => mskS))
 
-   𝑃=𝑃_Array2D{eltype(u0)}(u0[1], u1[1], v0[1], v1[1], (t0,t1))
+   𝑃=𝑃_Array2D{eltype(u0)}(u0[1], u1[1], v0[1], v1[1], [t0,t1])
    return 𝑃,Γ
 end
