+++ src/data_wrangling.jl
@@ -5,7 +5,7 @@
 Copy `sol` to a `DataFrame` & map position to lon,lat coordinates
 using "exchanged" 𝑃.XC, 𝑃.YC via `add_lonlat!`
 """
-function postprocess_lonlat(sol,𝑃::NamedTuple; id=missing, 𝑇=missing)
+function postprocess_lonlat(sol,𝑃::Union{NamedTuple,FlowParameters}; id=missing, 𝑇=missing)
     ismissing(id) ? id=collect(1:size(sol,2)) : nothing
     ismissing(𝑇) ? 𝑇=𝑃.𝑇 : nothing
 
@@ -122,22 +122,25 @@ function set_up_𝑃(k::Int,t::Float64,Γ::Dict,pth::String)
     iDYC=1. ./Γ["DYC"]
     γ=Γ["XC"].grid
     mon=86400.0*365.0/12.0
-
-    𝑃 = (u0=MeshArray(γ,Float64), u1=MeshArray(γ,Float64),
-         v0=MeshArray(γ,Float64), v1=MeshArray(γ,Float64),
-         𝑇=[-mon/2,mon/2], 🔄 = update_𝑃!, pth=pth,
+    func=Γ["update_location!"]
+    
+    𝐷 = (🔄 = update_𝑃!, pth=pth,
          XC=XC, YC=YC, iDXC=iDXC, iDYC=iDYC,
          k=k, msk=Γ["hFacC"][:, k])
 
     tmp = dict_to_nt(IndividualDisplacements.NeighborTileIndices_cs(Γ))
-    𝑃 = merge(𝑃 , tmp)
+    𝐷 = merge(𝐷 , tmp)
 
-    𝑃.🔄(𝑃,0.0)
-    return 𝑃
+    𝑃=𝑃_MeshArray2D{Float64}(MeshArray(γ,Float64),MeshArray(γ,Float64),
+    MeshArray(γ,Float64),MeshArray(γ,Float64),[-mon/2,mon/2],func)
+
+    𝐷.🔄(𝑃,𝐷,0.0)
+
+    return 𝑃,𝐷
 end
 
 """
-    update_𝑃!(𝑃::NamedTuple,t::Float64)
+    update_𝑃!(𝑃::Union{NamedTuple,FlowParameters},t::Float64)
 
 Update input data (velocity arrays) and time period (array) inside 𝑃 (𝑃.u0[:], etc, and 𝑃.𝑇[:])
 based on the chosen time `t` (in `seconds`). 
@@ -145,7 +148,7 @@ based on the chosen time `t` (in `seconds`).
 _Note: for now, it is assumed that (1) input 𝑃.𝑇 is used to infer `dt` between consecutive velocity fields,
 (2) periodicity of 12 monthly records, (3) vertical 𝑃.k is selected -- but this could easily be generalized._ 
 """
-function update_𝑃!(𝑃::NamedTuple,t::Float64)
+function update_𝑃!(𝑃::Union{NamedTuple,FlowParameters},𝐷::NamedTuple,t::Float64)
     dt=𝑃.𝑇[2]-𝑃.𝑇[1]
 
     m0=Int(floor((t+dt/2.0)/dt))
@@ -158,16 +161,16 @@ function update_𝑃!(𝑃::NamedTuple,t::Float64)
     m1=mod(m1,12)
     m1==0 ? m1=12 : nothing
 
-    (U,V)=read_velocities(𝑃.u0.grid,m0,𝑃.pth)
-    u0=U[:,𝑃.k]; v0=V[:,𝑃.k]
[…]
