### 2020-12-18 35a194a
new FlowParameters struct [4/n]
