+++ examples/helper_functions.jl
@@ -72,13 +72,15 @@ end
 function setup_point_cloud(U::Array{T,2},V::Array{T,2};X=[],Y=[]) where T
     np,nq=size(U)
     Γ=simple_periodic_domain(np,nq)
-    u=MeshArray(Γ["XC"].grid,[U])
-    v=MeshArray(Γ["XC"].grid,[V])
+    g=Γ["XC"].grid
+    u=MeshArray(g,[U])
+    v=MeshArray(g,[V])
     #vel=dxy_dt
     (u,v)=exchange(u,v,1)
     vel=dxy_dt!
+    func=(u -> IndividualDisplacements.update_location_dpdo!(u,g))
 
-    𝑃 = (u0=u, u1=u, v0=v, v1=v, 𝑇=[0.0,1.0], ioSize=Γ["XC"].grid.ioSize)
+    𝑃=𝑃_MeshArray2D{eltype(u)}(u,u,v,v,[0.0,10.0],func)
     pp=postprocess_xy
     isempty(X) ? X=np*rand(10) : nothing
     isempty(Y) ? Y=nq*rand(10) : nothing
@@ -114,14 +116,17 @@ function setup_global_ocean(;k=1,ny=2)
   Γ=GridLoad(γ)
   Γ=merge(Γ,IndividualDisplacements.NeighborTileIndices_cs(Γ))
 
+  func=(u -> IndividualDisplacements.update_location_llc!(u,𝐷))
+  Γ=merge(Γ,Dict("update_location!" => func))
+
   #initialize u0,u1 etc
-  𝑃=set_up_𝑃(k,0.0,Γ,ECCOclim_path);
+  𝑃,𝐷=set_up_𝑃(k,0.0,Γ,ECCOclim_path);
 
   #add parameters for use in reset!
   tmp=(frac=r_reset, Γ=Γ)
-  𝑃=merge(𝑃,tmp)
+  𝐷=merge(𝐷,tmp)
 
-  return 𝑃
+  return 𝑃,𝐷
 
 end
 
@@ -204,8 +209,12 @@ function OCCA_setup(;backward_in_time::Bool=false)
    n=length(Γ["RC"])
    n=10
 
+   g=Γ["XC"].grid
+   func=(u -> IndividualDisplacements.update_location_dpdo!(u,g))
+
    delete!.(Ref(Γ), ["hFacC", "hFacW", "hFacS","DXG","DYG","RAC","RAZ","RAS"]);
    backward_in_time ? s=-1.0 : s=1.0
+   s=Float32(s)
 
    function rd(filename, varname,n)
    fil = NetCDF.open(filename, varname)
@@ -273,12 +282,12 @@ function OCCA_setup(;backward_in_time::Bool=false)
     w[:,k]=tmpw
    end
 
-   𝑃 = (θ0=θ, θ1=θ, u0=u, u1=u, v0=v, v1=v, w0=w, w1=w, 𝑇=[t0,t1],
-   XC=exchange(Γ["XC"]), YC=exchange(Γ["YC"]), 
-   RF=Γ["RF"], RC=Γ["RC"],
-   ioSize=(360,160,n))
+   𝑃=𝑃_MeshArray3D{eltype(u)}(u,u,v,v,w,w,[t0,t1],func)
 
-   return 𝑃,Γ
+   𝐷 = (θ0=θ, θ1=θ, XC=exchange(Γ["XC"]), YC=exchange(Γ["YC"]), 
+   RF=Γ["RF"], RC=Γ["RC"],ioSize=(360,160,n))
+
+   return 𝑃,𝐷,Γ
 
 end
 
@@ -316,11 +325,11 @@ end
 
[…]
