### 2020-12-18 7bea74f
new FlowParameters struct [4/n]
Former-commit-id: 35a194a8045319f8a1fc6de500e902c9a1586885
