### 2020-12-19 JuliaClimate/IndividualDisplacements.jl#38 @visr
I'm running into the same issue wanting to apply this on a simple example grid of my own. I think the suggested syntax by @milankl could help here.

It's great to have many examples in the docs, but for me it seems the first tutorial (single particle) is too complicated. It would be best if it can be self explanatory, with no code being included from separate files. How about something super simple like this?

```julia
using Plots

X = range(0, 2π, length=60)
Y = range(0, π, length=30)
f(x, y) = sin(x) + cos(y)
ϕ = [f(x, y) for y in Y, x in X]
u = [cos(x) for y in Y, x in X]
v = [sin(y) for y in Y, x in X]

heatmap(X, Y, ϕ)
```
![image](https://user-images.githubusercontent.com/4471859/102698106-33883400-423b-11eb-99b6-9cf6c7900fec.png)

This would be a static 2D example. A next step could be adding a phase to show how this would work over time.
