### 2020-12-19 9e1b71b
fix a few typos
