### 2020-12-19 #2813 @visr
I'm basically done with review, only I don't yet feel confident to tick off the "Functionality" and "Functionality Documentation" boxes just yet, but that is being discussed in https://github.com/JuliaClimate/IndividualDisplacements.jl/issues/38.
