+++ test/helper_functions.jl
@@ -24,7 +24,7 @@ function test1_setup()
     u0=u./dx; u1=u./dx
     v0=v./dx; v1=v./dx
 
-    𝑃=𝑃_Array2D{eltype(u)}(u0[1], u1[1], v0[1], v1[1], [t0,t1])
+    𝑃=𝐹_Array2D{eltype(u)}(u0[1], u1[1], v0[1], v1[1], [t0,t1])
     
     u0=[200000.0;0.0]./dx
     du=fill(0.0,2);
@@ -60,7 +60,7 @@ function test2_periodic_domain(np = 12, nq = 12)
     (u, v) = exchange(u, v, 1)
 
     f = (u -> IndividualDisplacements.update_location_dpdo!(u,Γ.XC.grid))
-    𝑃=𝑃_MeshArray2D{eltype(u)}(u,u,v,v,[0.0,400.0],f)
+    𝑃=𝐹_MeshArray2D{eltype(u)}(u,u,v,v,[0.0,400.0],f)
 
     #initial conditions
     x0 = np * (0.4:0.04:0.6)
