+++ examples/helper_functions.jl
@@ -80,7 +80,7 @@ function setup_point_cloud(U::Array{T,2},V::Array{T,2};X=[],Y=[]) where T
     vel=dxy_dt!
     func=(u -> IndividualDisplacements.update_location_dpdo!(u,g))
 
-    𝑃=𝑃_MeshArray2D{eltype(u)}(u,u,v,v,[0.0,10.0],func)
+    𝑃=𝐹_MeshArray2D{eltype(u)}(u,u,v,v,[0.0,10.0],func)
     pp=postprocess_xy
     isempty(X) ? X=np*rand(10) : nothing
     isempty(Y) ? Y=nq*rand(10) : nothing
@@ -282,7 +282,7 @@ function OCCA_setup(;backward_in_time::Bool=false)
     w[:,k]=tmpw
    end
 
-   𝑃=𝑃_MeshArray3D{eltype(u)}(u,u,v,v,w,w,[t0,t1],func)
+   𝑃=𝐹_MeshArray3D{eltype(u)}(u,u,v,v,w,w,[t0,t1],func)
 
    𝐷 = (θ0=θ, θ1=θ, XC=exchange(Γ["XC"]), YC=exchange(Γ["YC"]), 
    RF=Γ["RF"], RC=Γ["RC"],ioSize=(360,160,n))
