### 2020-12-20 8f6b4bf
update worklow docs -> new Individuals(𝐹::��_Array3D,x,y,z); update solid_body_rotation
Former-commit-id: 332a0c10f3b2865e65c41572e27755543d772038 [formerly e4da9b6c1968f89427c6aeb2948b31a963043284]
Former-commit-id: 9d4b302f180823bfca5f03b74bf1b996b98915a9
Former-commit-id: ea90890db56b72f2c869c0443ad712359ff1ad8c
