+++ Project.toml
@@ -1,7 +1,7 @@
 name = "IndividualDisplacements"
 uuid = "b92f0c32-5b7e-11e9-1d7b-238b2da8b0e6"
 authors = ["gaelforget <gforget@mit.edu>"]
-version = "0.2.6"
+version = "0.2.7"
 
 [deps]
 CFTime = "179af706-886a-5703-950a-314cd64e0468"
