+++ examples/worldwide/global_ocean_circulation.jl
@@ -45,16 +45,7 @@ fieldnames(typeof(𝑃))
 # - initial particle positions randomly over Global Ocean
 
 xy = init_global_randn(1000,𝐷)
-xy = permutedims([xy[:,i] for i in 1:size(xy,2)])
-
-function 🔧(sol,𝑃::𝐹_MeshArray2D;id=missing,𝑇=missing)
-    df=postprocess_lonlat(sol,𝐷,id=id,𝑇=𝑇)
-end
- 
-I=(position=xy,velocity=dxy_dt!,
-   postprocessing=🔧,parameters=𝑃)
-𝐼=Individuals(I)
-
+𝐼=Individuals(𝑃,xy[1,:],xy[2,:],xy[3,:])
 fieldnames(typeof(𝐼))
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
@@ -62,34 +53,45 @@ fieldnames(typeof(𝐼))
 
 𝑇=(0.0,𝐼.𝑃.𝑇[2])
 ∫!(𝐼,𝑇)
+add_lonlat!(𝐼.🔴,𝐷.XC,𝐷.YC);
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
 # ### 3.2 Iteration function example
 #
-# - `𝐼.𝑃.🔄(𝐼.𝑃,t_ϵ)` resets the velocity input streams to bracket t_ϵ=𝐼.𝑃.𝑇[2]+eps(𝐼.𝑃.𝑇[2]) 
+# Here we customize the postprocessing function to add longitude and latitude (my🔧).
+
+function my🔧(sol,𝑃::𝐹_MeshArray2D;id=missing,𝑇=missing)
+    df=postprocess_MeshArray(sol,𝑃,id=id,𝑇=𝑇)
+    add_lonlat!(df,𝐷.XC,𝐷.YC)
+end
+
+𝐽=Individuals{eltype(𝐼.📌),ndims(𝐼.📌)}(𝑃=𝐼.𝑃,📌=𝐼.📌,🔴=𝐼.🔴,🆔=𝐼.🆔,🚄=𝐼.🚄,∫=𝐼.∫,🔧=my🔧)
+
+# In addition, `step!` is defined to provide additional flexibility around `∫!` :
+#
+# - `𝐷.🔄(𝐼.𝑃,t_ϵ)` resets the velocity input streams to bracket t_ϵ=𝐼.𝑃.𝑇[2]+eps(𝐼.𝑃.𝑇[2]) 
 # - `reset_lonlat!(𝐼)` randomly selects a fraction (defined in `setup_global_ocean()`) of the particles and resets their positions before each integration period. This can maintain homogeneous coverage of the Global Ocean by particles.
-# - `∫!(𝐼,𝑇)` then solves for the individual trajectories over one month, after updating velocity fields (𝐼.u0 etc) if needed, and adds diagnostics to the DataFrame used to record / trace variables along the trajectory (𝐼.tr).
+# - `∫!(𝐼)` then solves for the individual trajectories over one month, after updating velocity fields (𝐼.u0 etc) if needed, and adds diagnostics to the DataFrame used to record / trace variables along the trajectory (𝐼.tr).
 
 function step!(𝐼::Individuals)
     t_ϵ=𝐼.𝑃.𝑇[2]+eps(𝐼.𝑃.𝑇[2])
     𝐷.🔄(𝐼.𝑃,𝐷,t_ϵ)
     reset_lonlat!(𝐼,𝐷)
-    𝑇=Tuple(𝐼.𝑃.𝑇)
-    ∫!(𝐼,𝑇)
+    ∫!(𝐼)
 end
[…]
