### 2020-12-21 7dfe59b
update docs and add missing constructors
