### 2020-12-21 3d660d2
update docs and add missing constructors
Former-commit-id: 1c2f79924e738f316203cb076f4a39867a65ef4f [formerly 7dfe59bd3146f4010f1d3744f41e3e99215b9b4a]
Former-commit-id: 95182fb1c4c0b7b7b6dabd98d58b592a0abc90b1
Former-commit-id: d67eaa0f89363555bc1f1eaac7e5bc8330db98bb
