### 2020-12-22 ac7f7d3
reduce allocations, improve type stability
Former-commit-id: ceb240d205ada0a2321e728ce964bec4e40e7c00
