### 2020-12-22 62da73f
less strict type annotation + fix in three_dimensional_ocean
