### 2020-12-22 ad887d2
less strict type annotation + fix in three_dimensional_ocean
Former-commit-id: d330eda8a225f890905d6ce769002eb25aad801c [formerly a9a874903dbcce1aa5146e67a3ccab5c8a0e3abb]
Former-commit-id: 66cb038eda347d5d9e9de8b0554c666c7db5be06
Former-commit-id: 58d303e005836c4f7408921f3a1e01da06d75b3f
