### 2021-01-22 5781b23
improve docs, defaults handling + minor fixes
Former-commit-id: 06cc4c18a6c6b49b5e086847233d04b8710bdfc4 [formerly 50b6a32d36638087a7819b9efc18e7d26c613c17]
Former-commit-id: c1f409e4538656734bab98c09b3330a9058d722b
Former-commit-id: bb22cb8341d68ba6ab2bcdb63826555323ddab5e
