### 2021-01-22 06cc4c1
improve docs, defaults handling + minor fixes
Former-commit-id: 50b6a32d36638087a7819b9efc18e7d26c613c17
