### 2021-01-22 50b6a32
improve docs, defaults handling + minor fixes
