+++ src/API.jl
@@ -68,16 +68,11 @@ end
     defaults for Individuals constructor
 """
 
-day=86400.0
-mon=365/12*day
-OneMonth=[-0.5*mon,0.5*mon]
-
-solver_default(prob) = solve(prob,Tsit5(),reltol=1e-8,abstol=1e-8)
-#solver_default(prob) = solve(prob,Euler(),dt=day)
-param_default = ( 𝑇=OneMonth , 🔄=(x->x), u0=[], u1=[], v0=[], v1=[])
-rec_default = DataFrame(ID=Int[], x=Float64[], y=Float64[], t=Float64[], 
-                        lon=Float64[], lat=Float64[], fid=Int[])
-postprocess_default = (x->x)
+default_solver(prob) = solve(prob,Tsit5(),reltol=1e-8,abstol=1e-8)
+a=fill(0.0,1,1)
+default_flowfields = 𝐹_Array2D{Float64}(a,a,a,a,[0. 1.])
+default_recorder = DataFrame(ID=Int[], x=Float64[], y=Float64[], t=Float64[])
+default_postproc = (x->x)
 
 """
     struct Individuals{T}
@@ -86,10 +81,23 @@ postprocess_default = (x->x)
 - Functions:      🚄 (velocity),   ∫ (integration), 🔧(postprocessing)
 - NamedTuples:    𝐷 (diagnostics),      𝑀 (metadata)
 
-The velocity function 🚄 typically computes velocity at an arbitrary position within the 
-chosen space-time domain (📌 to start) by interpolating gridded variables obtained from 𝑃.
+The velocity function 🚄 typically computes velocity at individual positions (📌 to start) within the 
+specified space-time domain by interpolating gridded variables (provided via 𝑃). Individual trajectories 
+are computed by integrating (∫) interpolated velocities through time. Normally, integration is done by 
+calling ∫! which updates 📌 at the end and records results in 🔴 via 🔧. Unicode cheatsheet:
 
-Default keyword constructor example:
+- 📌=`\\:pushpin:<tab>`,          🔴=`\\:red_circle:<tab>`, 🆔=`\\:id:<tab>`
+- 🚄=`\\:bullettrain_side:<tab>`, ∫=`\\int<tab>`,          🔧=`\\wrench<tab>`
+- 𝑃=`\\itP<tab>`,                 𝐷=`\\itD<tab>`,           𝑀=`\\itM<tab>`
+
+Simple constructors that use `FlowFields` to choose adequate defaults:
+
+- Individuals(𝐹::𝐹_Array2D,x,y)
+- Individuals(𝐹::𝐹_Array3D,x,y,z)
+- Individuals(𝐹::𝐹_MeshArray2D,x,y,fid)
+- Individuals(𝐹::𝐹_MeshArray3D,x,y,z,fid)
+
+Further customization is achievable via keyword constructors:
 
 ```
 df=DataFrame( ID=[], x=[], y=[], z=[], t = [])
@@ -97,52 +105,35 @@ df=DataFrame( ID=[], x=[], y=[], z=[], t = [])
 𝐼=Individuals(📌=zeros(3,2),🆔=collect(1:2),🔴=deepcopy(df))
[…]
