### 2021-01-22 #2813 @gaelforget
> @milankl ah ok, got it.
>
> @gaelforget What is your time frame for working out the remaining review issues? I can pause this issue if you'd like more than a couple of weeks (assuming the reviewers agree).

Hi @kthyng

Am working on it now and hope to be done early next week.

As you saw, I had closed the other, also very helpful, issues raised by @milankl and @visr in Dec. Sorry for the long break afterwards. I resumed with a bit of cleanup on the code and docs (https://github.com/JuliaClimate/IndividualDisplacements.jl/pull/52) which should help close https://github.com/JuliaClimate/IndividualDisplacements.jl/issues/38 and once done I would release v0.2.8 , merge  into there paper branch, add a summary of functionalities, and post responses on the issue tracker if that works.

Thank you all for your patience and contributions to this review so far
