### 2021-01-23 1c05f05
gather flow field setup examples into examples/flow_fields.jl -- also: - simpler names for flow field example functions - drop example3_setup (redundant with OCCA) - rm old_setup_point_cloud - reorder in docs/src/workflow.md
Former-commit-id: b513c4e7fd97ad34e0f842d72f6306c1a798fd8b
