+++ examples/flow_fields.jl
@@ -1,0 +1,274 @@
+
+"""
+    random_flow_field(;np=12,nq=18)
+
+Set up a random flow field over a gridded domain of size np,nq
+
+```
+ϕ,u,v=random_flow_field()
+```
+"""
+function random_flow_field(;np=12,nq=18)
+Γ=simple_periodic_domain(np,nq)
+(_,ϕ,_,_)=demo2(Γ)
+ϕ .*= 0.5
+
+#For the convergent / scalar potential case, ϕ is interpreted as being 
+#on center points -- hence the standard gradient function readily gives 
+#what we need
+#(u,v)=gradient(ϕ,Γ) 
+#return u[1],v[1],ϕ[1]
+
+#For the rotational / streamfunction case, ϕ is interpreted as being 
+#on S/W corner points -- this is ok since the grid is homegeneous, 
+#and conveniently yields an adequate substitution u,v <- -v,u; but note
+#that doing the same with gradient() would shift indices inconsistenly
+u=-(circshift(ϕ[1], (0,-1))-ϕ[1])
+v=(circshift(ϕ[1], (-1,0))-ϕ[1])
+return u,v,ϕ[1]
+end
+
+"""
+    solid_body_rotation(np,nz)
+
+Set up an idealized flow field which consists of 
+[rigid body rotation](https://en.wikipedia.org/wiki/Rigid_body), 
+plus a convergent term, plus a sinking term.
+
+```
+u,v,w=solid_body_rotation(12,4)
+```
+"""
+function solid_body_rotation(np,nz)
+    Γ=simple_periodic_domain(np);
+    γ=Γ["XC"].grid;
+    
+    #Solid-body rotation around central location ...
+    i=Int(np/2+1)
+    u=-(Γ["YG"].-Γ["YG"][1][i,i])
+    v=(Γ["XG"].-Γ["XG"][1][i,i])
+    
+    #... plus a convergent term to / from central location
+    d=-0.01
+    u=u+d*(Γ["XG"].-Γ["XG"][1][i,i])
+    v=v+d*(Γ["YG"].-Γ["YG"][1][i,i])
+    
+    #Replicate u,v in vertical dimension
+    uu=MeshArray(γ,γ.ioPrec,nz)
+    [uu[k]=u[1] for k=1:nz]
+    vv=MeshArray(γ,γ.ioPrec,nz)
+    [vv[k]=v[1] for k=1:nz]
+    
+    #Vertical velocity component w    
+    w=fill(-0.01,MeshArray(γ,γ.ioPrec,nz));
+    
+    return write(uu),write(vv),write(w)
+end
+
+"""
+    global_ocean_circulation(;k=10,ny=2)
+
+Set up Global Ocean particle simulation in 2D with seasonally varying flow field.
+
+```
+𝑃,𝐷=global_ocean_circulation(k=10,ny=2);
+```
+"""
+function global_ocean_circulation(;k=1,ny=2)
+
+  #k=10 #choice of vertical level
+  #ny=2 #number of simulated years (20 for k>20)
+  r_reset = 0.01 #fraction of the particles reset per month (0.05 for k<=10)
+
+  #read grid and set up connections between subdomains
[…]
