### 2021-01-23 da60033
gather flow field setup examples into examples/flow_fields.jl -- also: - simpler names for flow field example functions - drop example3_setup (redundant with OCCA) - rm old_setup_point_cloud - reorder in docs/src/workflow.md
Former-commit-id: 1c05f057993b6055cea6965ed8062f948a3b242d [formerly b513c4e7fd97ad34e0f842d72f6306c1a798fd8b]
Former-commit-id: 1d8c651e0e7e1ad594e21cda4a54625c3652fcd8
Former-commit-id: 7bac70505819908328804792f2a1b874dd65bbf5
