+++ examples/helper_functions.jl
@@ -2,300 +2,31 @@
 using IndividualDisplacements, MeshArrays, OrdinaryDiffEq
 p=dirname(pathof(MeshArrays))
 include(joinpath(p,"../examples/Demos.jl"))
-
-"""
-    simple_flow_field(Γ::Dict,np,nz)
-
-Set up an idealized flow field which consists of 
-[rigid body rotation](https://en.wikipedia.org/wiki/Rigid_body), 
-plus a convergent term, plus a sinking term.
-
-```
-u,v,w=simple_flow_field(12,4)
-```
-"""
-function simple_flow_field(np,nz)
-    Γ=simple_periodic_domain(np);
-    γ=Γ["XC"].grid;
-    
-    #Solid-body rotation around central location ...
-    i=Int(np/2+1)
-    u=-(Γ["YG"].-Γ["YG"][1][i,i])
-    v=(Γ["XG"].-Γ["XG"][1][i,i])
-    
-    #... plus a convergent term to / from central location
-    d=-0.01
-    u=u+d*(Γ["XG"].-Γ["XG"][1][i,i])
-    v=v+d*(Γ["YG"].-Γ["YG"][1][i,i])
-    
-    #Replicate u,v in vertical dimension
-    uu=MeshArray(γ,γ.ioPrec,nz)
-    [uu[k]=u[1] for k=1:nz]
-    vv=MeshArray(γ,γ.ioPrec,nz)
-    [vv[k]=v[1] for k=1:nz]
-    
-    #Vertical velocity component w    
-    w=fill(-0.01,MeshArray(γ,γ.ioPrec,nz));
-    
-    return write(uu),write(vv),write(w)
-end
+p=dirname(pathof(IndividualDisplacements))
+include(joinpath(p,"../examples/flow_fields.jl"));
     
 """
-    setup_random_flow(;np=12,nq=18)
+    convert_to_FlowFields(U::Array{T,2},V::Array{T,2}) where T
 
-Set up a random flow field over a gridded domain of size np,nq
+Convert a pair of U,V arrays (staggered C-grid velocity field in 2D) to
+a `𝐹_MeshArray2D` struct ready for integration of individual displacements.
 
 ```
-ϕ,u,v=setup_random_flow()
+_,u,v=random_flow_field()
+𝐹=convert_to_FlowFields(u,v)
 ```
 """
-function setup_random_flow(;np=12,nq=18)
-    Γ=simple_periodic_domain(np,nq)
-    (_,ϕ,_,_)=demo2(Γ)
-    ϕ .*= 0.5
-
-    #For the convergent / scalar potential case, ϕ is interpreted as being 
-    #on center points -- hence the standard gradient function readily gives 
-    #what we need
-    #(u,v)=gradient(ϕ,Γ) 
-    #return u[1],v[1],ϕ[1]
-
-    #For the rotational / streamfunction case, ϕ is interpreted as being 
-    #on S/W corner points -- this is ok since the grid is homegeneous, 
-    #and conveniently yields an adequate substitution u,v <- -v,u; but note
-    #that doing the same with gradient() would shift indices inconsistenly
-    u=-(circshift(ϕ[1], (0,-1))-ϕ[1])
[…]
