### 2021-01-23 b513c4e
gather flow field setup examples into examples/flow_fields.jl -- also: - simpler names for flow field example functions - drop example3_setup (redundant with OCCA) - rm old_setup_point_cloud - reorder in docs/src/workflow.md
