### 2021-01-23 e18d73a
improved plotting recipes
Former-commit-id: 66464492eb9cfed423156449fc65df31faa56fe5
