### 2021-01-25 f4ee721
CompatHelper: bump compat for "CyclicArrays" to "0.5"
Former-commit-id: a70920ec4c01181cd232fedac665201d69ce9ff9
