+++ Project.toml
@@ -21,7 +21,7 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 [compat]
 CFTime = "0.1"
 CSV = "0.6, 0.7, 0.8"
-CyclicArrays = "0.2, 0.3, 0.4"
+CyclicArrays = "0.2, 0.3, 0.4, 0.5"
 DataFrames = "0.21, 0.22"
 MITgcmTools = "0.1.6"
 MeshArrays = "0.2.14"
