### 2021-01-25 aa2d2c9
streamline and update example
