+++ examples/worldwide/three_dimensional_ocean.jl
@@ -30,33 +30,15 @@ IndividualDisplacements.get_occa_velocity_if_needed();
 # ## 2.1 Ocean Circulation Setup
 #
 
-𝑃,𝐷,Γ=OCCA_FlowFields(backward_in_time=false)
+𝑃,𝐷,Γ=OCCA_FlowFields(backward_in_time=false);
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
-# ## 2.2 Solver And Analysis Setup
+# ## 2.2 Post-Processor Setup
 #
 
-function 🚄(du::Array{T,2},u::Array{T,2},𝑃::𝐹_MeshArray3D,tim) where T
-   nf=size(u,2)
-   nx=360
-   ny=160
-   [u[1,i][1]=mod(u[1,i][1],nx) for i in 1:nf]
-   [u[1,i][2]=mod(u[1,i][2],ny) for i in 1:nf]
-   [dxyz_dt!(du[i],u[i],𝑃,tim) for i=1:size(u,2)]
-end
-
-function ∫(prob)
-   #sol=solve(prob,Tsit5(),reltol=1e-8,abstol=1e-8,saveat=5*86400.0)
-   sol=IndividualDisplacements.default_solver(prob)
-   #sol=solve(prob,Euler(),dt=86400.0)
-
-   nx,ny=𝑃.u0.grid.ioSize[1:2]
-   nf=size(sol,2)
-   nt=size(sol,3)
-   [sol[1,i,j][1]=mod(sol[1,i,j][1],nx) for i in 1:nf, j in 1:nt]
-   [sol[1,i,j][2]=mod(sol[1,i,j][2],ny) for i in 1:nf, j in 1:nt]
-   return sol
-end
+tr = DataFrame(ID=Int[], fid=Int[], x=Float64[], y=Float64[], 
+               k=Float64[], z=Float64[], iso=Float64[], t=Float64[], 
+               lon=Float64[], lat=Float64[], year=Float64[], col=Symbol[])
 
 function 🔧(sol,𝑃::𝐹_MeshArray3D;id=missing,𝑇=missing)
    df=postprocess_MeshArray(sol,𝑃,id=id,𝑇=𝑇)
@@ -89,46 +71,54 @@ function 🔧(sol,𝑃::𝐹_MeshArray3D;id=missing,𝑇=missing)
 end
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
-# ## 2.3 Initialize Individuals
+# ## 2.3 Initialize Individual Positions
 #
 
 """
-    set_up_individuals(𝑃,Γ,∫,🚄,🔧; nf=10000, z_init=4.5, 
-               lon_rng=(-160.0,-150.0), lat_rng=(30.0,40.0))
+    initial_positions(Γ; nf=10000, lon_rng=(-160.0,-159.0), lat_rng=(30.0,31.0))
 
-Set up `Individuals` data structure with `nf` particles moving within a near-global Ocean domain. 
+Randomly assign initial positions in longitude,latitude ranges. Positions are 
+expressed in, normalized, grid point units (x,y in the 0,nx and 0,ny range). 
+To convert from longitude,latitude here we take advantage of the regularity 
+of the 1 degree grid being used -- for a more general alternative, see 
+`initialize_lonlat()`.
 """
-function set_up_individuals(𝑃,Γ,∫,🚄,🔧; nf=10000, 
[…]
