+++ examples/basics/random_flow_field.jl
@@ -24,15 +24,20 @@
 
 using IndividualDisplacements, DataFrames
 p=dirname(pathof(IndividualDisplacements))
-include(joinpath(p,"../examples/helper_functions.jl"));
+include(joinpath(p,"../examples/flow_fields.jl"));
 
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ## 2. Flow Fields
 #
+# A convenient way to set up the flow fields using the MeshArrays.jl package (which 
+# handles such staggered grids in general fashion) is to call the `convert_to_FlowFields`
+# function with `u,v` arrays and time span as arguments.
 
-u,v,ϕ=random_flow_field();
+u,v,ϕ=random_flow_field()
 
-# The above `u,v` arrays can be replaced with any other pair provided by the user.
+𝐹=convert_to_FlowFields(u,v,10.0);
+
+# In the above, the `u,v` arrays generated by `random_flow_field` can be replaced with any other pair provided by the user.
 #
 # A couple of important considerations, however:
 #
@@ -55,12 +60,6 @@ u,v,ϕ=random_flow_field();
 # u=0.5*(circshift(uC, (0,1))+uC)
 # v=0.5*(circshift(vC, (1,0))+vC)
 # ```
-
-# A convenient way to set up the flow fields using the MeshArrays.jl package (which 
-# handles such staggered grids in general fashion) is to call the `convert_to_FlowFields`
-# function with `u,v` arrays and time span as arguments.
-
-𝐹=convert_to_FlowFields(u,v,10.0);
 
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ## 3. Initialize Individuals
