+++ src/IndividualDisplacements.jl
@@ -15,7 +15,7 @@ include("data_wrangling.jl")
 include("read.jl")
 include("update_locations.jl")
 
-export Individuals, ∫!, set_up_𝑃, update_𝑃!
+export Individuals, ∫!, set_up_𝑃, update_𝑃!, convert_to_FlowFields
 export FlowFields, 𝐹_Array3D, 𝐹_Array2D, 𝐹_MeshArray3D, 𝐹_MeshArray2D
 export dxy_dt!, dxy_dt, dxyz_dt!, dxyz_dt, dxy_dt_CyclicArray, dxy_dt_replay
 export postprocess_MeshArray, add_lonlat!, postprocess_xy, interp_to_xy
