### 2021-01-26 dc0409e
add initial position files + simplify example
Former-commit-id: 3ff01904a89e99487edb81ca2550c17e681e20f5
