### 2021-01-26 bde5805
reduce memory footprint of example
