### 2021-01-26 46c5f6d
reduce memory footprint of example
Former-commit-id: 3ecf0cbbd3099c476a595f34a8ff51b7935a59d2 [formerly bde580594c2f443a82bc11e5d35ea3f7fbed0006]
Former-commit-id: 5cc3a842f9d8f825e1190b1eb435a17fdaa65624
Former-commit-id: 7359aae15158c0800031c5b6c40dcc08a1d50a25
