### 2021-01-26 12342ca
udpate / clean up recipes_plots
Former-commit-id: 1f5567a1c3b6be8987b546c113864960e9566244 [formerly ce2f9d12a39c2ed9a26908b152ed192e5d4fbea3]
Former-commit-id: 4072da89ac91b390c0092ee2a65f9391e4a16528
Former-commit-id: f463fc6668e2a8b48cf3d3e1484ea1675eb7af64
