+++ examples/worldwide/three_dimensional_ocean.jl
@@ -11,8 +11,7 @@
 # [3](https://docs.juliadiffeq.org/latest/solvers/ode_solve.html),
 # [4](https://en.wikipedia.org/wiki/Displacement_(vector))
 #
-# ![Three dimensional simulation 1/2](https://user-images.githubusercontent.com/20276764/94491655-9cb95780-01b6-11eb-9273-173865ed6340.png)
-# ![Three dimensional simulation 2/2](https://user-images.githubusercontent.com/20276764/94491485-595ee900-01b6-11eb-95e6-c2cacb812f46.png)
+# ![Three dimensional simulation](https://user-images.githubusercontent.com/20276764/94491485-595ee900-01b6-11eb-95e6-c2cacb812f46.png)
 
 #nb # %% {"slideshow": {"slide_type": "subslide"}, "cell_type": "markdown"}
 # ## 1. Load Software
@@ -135,7 +134,7 @@ set_up_individuals(𝐼::Individuals; nf=10000) = set_up_individuals(𝑃,Γ,�
 
 #!jl #include(joinpath(p,"../examples/recipes_plots.jl"))
 #!jl #p=plot(𝐼)
-#!jl ##p=plot_end_points(𝐼,Γ)
+#!jl ##p=map(𝐼,OceanDepthLog(Γ))
 #!jl #display(p)
 
 # - or `Makie.jl`:
