### 2021-01-30 2c6b786
reorganize and revise docs
