### 2021-02-04 68a082d
simplify first example (use Array + flow field suggested by @visr)
