+++ examples/basics/random_flow_field.jl
@@ -29,15 +29,40 @@ include(joinpath(p,"../examples/flow_fields.jl"));
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ## 2. Flow Fields
 #
+
+nx=30
+dx= π/nx
+XC = dx*(collect(1:2*nx) .- 0.5)
+YC = dx*(collect(1:nx) .- 0.5)
+
+f(x, y) = sin(x) + cos(y) #streamfunction
+ϕ = [f(x, y) for x in XC,y in YC] #streamfunction
+uC = -[sin(y) for x in XC,y in YC] #dphi/dy at cell center
+vC = -[cos(x) for x in XC,y in YC]; #-dphi/dx at cell center
+
+# When starting in this fashion with collocated velocities in m/s units (`uC,vC` at the grid cell center), 
+# one easily obtains the normalized, staggered C-grid velocities (`u,v`) as shown below. Any uC, vC
+# computed as done above may contain contain both 
+# [rotational and divergent](https://en.wikipedia.org/wiki/Helmholtz_decomposition)
+# components -- if this were an important consideration, the MeshArrays.jl package
+# tools can be used to e.g. ensure non-divergence. Normalization of velocities 
+# by the grid scale is the convention in IndividualDisplacements.jl package
+
+u=0.5*(circshift(uC, (1,0))+uC)/dx #staggered u converted to grid point units (m/s -> 1/s)
+v=0.5*(circshift(vC, (0,1))+vC)/dx #staggered v converted to grid point units (m/s -> 1/s)
+
+𝐹=𝐹_Array2D(u,u,v,v,[0.,10.])
+
 # A convenient way to set up the flow fields using the MeshArrays.jl package (which 
 # handles such staggered grids in general fashion) is to call the `convert_to_FlowFields`
 # function with `u,v` arrays and time span as arguments.
+#
+# ```
+# u,v,ϕ=random_flow_field()
+# 𝐹=convert_to_FlowFields(u,v,100.0);
+# ```
 
-u,v,ϕ=random_flow_field()
-
-𝐹=convert_to_FlowFields(u,v,10.0);
-
-# In the above, the `u,v` arrays generated by `random_flow_field` can be replaced with any other pair provided by the user.
+# The `u,v` arrays generated by `random_flow_field` can be replaced with any other pair provided by the user.
 #
 # A couple of important considerations, however:
 #
@@ -45,20 +70,12 @@ u,v,ϕ=random_flow_field()
 #  from the grid cell center (0.5,0.5)
 # - `u,v` here derive from streamfunction `ϕ`, defined at the corner point, which ensures that 
 #  the resulting `u,v` is non-divergent, purely rotational, over the C-grid domain.
+# 
 # In brief:
 #
 # ```
 # u=-(circshift(ϕ, (0,-1))-ϕ)
 # v=(circshift(ϕ, (-1,0))-ϕ)
-# ```
-
-# If user were to start with collocated velocity (`uC,vC` at the grid cell center) then
-# one can easily obtain the staggered velocity (`u,v`) as follows. These may contain both 
-# [rotational and divergent](https://en.wikipedia.org/wiki/Helmholtz_decomposition) components.
-#
-# ```
-# u=0.5*(circshift(uC, (0,1))+uC)
-# v=0.5*(circshift(vC, (1,0))+vC)
 # ```
 
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
@@ -69,11 +86,13 @@ u,v,ϕ=random_flow_field()
[…]
