### 2021-02-04 b4e7737
simplify first example (use Array + flow field suggested by @visr)
Former-commit-id: 68a082d04b2fc96764c9ee327fc63a3f71dacc85
