### 2021-02-04 6b3a6c5
improve docs and bump version
Former-commit-id: bf05dc4374ef19a4b170372b4075f6c0d0ddad15
