+++ examples/basics/random_flow_field.jl
@@ -1,10 +1,10 @@
-# # Particle Set
+# # Two Dimensions
 #
 #md # [![](https://mybinder.org/badge_logo.svg)](@__BINDER_ROOT_URL__/notebooks/random_flow_field.ipynb)
 #md # [![](https://img.shields.io/badge/show-nbviewer-579ACA.svg)](@__NBVIEWER_ROOT_URL__/notebooks/random_flow_field.ipynb)
 #
 # Simulate trajectories of a particle cloud in a two-dimensional flow field.
-# A doubly-periodic domain and randomly-generated flow fields are initially used.
+# A doubly-periodic domain and simple flow fields are used. 
 # For additional documentation e.g. see :
 # [1](https://JuliaClimate.github.io/IndividualDisplacements.jl/dev/),
 # [2](https://JuliaClimate.github.io/MeshArrays.jl/dev/),
@@ -12,7 +12,7 @@
 # [4](https://en.wikipedia.org/wiki/Displacement_(vector))
 #
 # Exercises: 
-# - change the initial distribution of partices
+# - change the initial distribution of particles
 # - increase the duration of the trajectories simulation
 # - treat the non-periodic domain case by padding `u,v` with zeros 
 # - replace `u,v` with your own two-dimensional flow fields 
@@ -29,6 +29,9 @@ include(joinpath(p,"../examples/flow_fields.jl"));
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ## 2. Flow Fields
 #
+# Let's start with a simple, doubly periodic flow field defined by a streamfunction
+# and create the `FlowFields` data structure which will then be drive the 
+# individual displacement and trajectory computations
 
 nx=30
 dx= π/nx
@@ -40,36 +43,48 @@ f(x, y) = sin(x) + cos(y) #streamfunction
 uC = -[sin(y) for x in XC,y in YC] #dphi/dy at cell center
 vC = -[cos(x) for x in XC,y in YC]; #-dphi/dx at cell center
 
-# When starting in this fashion with collocated velocities in m/s units (`uC,vC` at the grid cell center), 
-# one easily obtains the normalized, staggered C-grid velocities (`u,v`) as shown below. Any uC, vC
-# computed as done above may contain contain both 
+# It should be noted that any uC, vC computed as done here may contain contain both 
 # [rotational and divergent](https://en.wikipedia.org/wiki/Helmholtz_decomposition)
-# components -- if this were an important consideration, the MeshArrays.jl package
-# tools can be used to e.g. ensure non-divergence. Normalization of velocities 
-# by the grid scale is the convention in IndividualDisplacements.jl package
+# components -- if this were an important consideration, `MeshArrays.jl` provides 
+# tools that can be used to e.g. ensure non-divergence. 
+#
+# When starting as done here with collocated velocities in m/s units (`uC,vC` at 
+# the grid cell center), one easily obtains the normalized, staggered C-grid 
+# velocities (`u,v`) as shown below. Both the staggering of variables and the 
+# normalization of velocities by the grid scale are key conventions
+# adopted in `IndividualDisplacements.jl`.
 
-u=0.5*(circshift(uC, (1,0))+uC)/dx #staggered u converted to grid point units (m/s -> 1/s)
-v=0.5*(circshift(vC, (0,1))+vC)/dx #staggered v converted to grid point units (m/s -> 1/s)
[…]
