### 2021-02-04 c23b221
improve docs and bump version
Former-commit-id: 6b3a6c559ec3097e6bb2118bc479ae2fe7663513 [formerly bf05dc4374ef19a4b170372b4075f6c0d0ddad15]
Former-commit-id: 9b9f24556aa2bedb96a200e330dda430bd1de83e
Former-commit-id: 25dbf8ba3e14483eb8c9c5fce8a399f80cda29c9
