### 2021-02-04 6a1fc99
fix typos noted by @visr (Thanks!)
