### 2021-02-04 9c2b411
fix typos noted by @visr (Thanks!)
Former-commit-id: 6a1fc99acdd887bbe7a3ad2dd4feb784380df5c5
