+++ src/API.jl
@@ -97,7 +97,7 @@ are computed by integrating (∫) interpolated velocities through time. Normally
 calling ∫! which updates 📌 at the end and records results in 🔴 via 🔧. Unicode cheatsheet:
 
 - 📌=`\\:pushpin:<tab>`,          🔴=`\\:red_circle:<tab>`, 🆔=`\\:id:<tab>`
-- 🚄=`\\:bullettrain_side:<tab>`, ∫=`\\int<tab>`,          🔧=`\\wrench<tab>`
+- 🚄=`\\:bullettrain_side:<tab>`, ∫=`\\int<tab>`,          🔧=`\\:wrench:<tab>`
 - 𝑃=`\\itP<tab>`,                 𝐷=`\\itD<tab>`,           𝑀=`\\itM<tab>`
 
 Simple constructors that use `FlowFields` to choose adequate defaults:
@@ -129,10 +129,10 @@ Base.@kwdef struct Individuals{T,N}
    🆔   ::Array{Int,1} = Array{Int,1}(undef, 0) #\:id:<tab>
    🚄  ::Function = dxy_dt #\:bullettrain_side:<tab>
    ∫   ::Function = default_solver #\int<tab>
-   🔧  ::Function = default_postproc #\wrench<tab>
+   🔧  ::Function = default_postproc #\:wrench:<tab>
    𝑃   ::FlowFields = default_flowfields #\itP<tab>
    𝐷   ::NamedTuple = NamedTuple() #\itD<tab>
-   𝑀   ::NamedTuple = NamedTuple() #\itM<tab>vec
+   𝑀   ::NamedTuple = NamedTuple() #\itM<tab>
 end
 
 function Individuals(NT::NamedTuple)
@@ -220,7 +220,7 @@ Displace simulated individuals continuously through space over time period 𝑇 
 
 - This is typically achieved by computing the cumulative integral of velocity experienced by each individual along its trajectory (∫ 🚄 dt).
 - The current default is `solve(prob,Tsit5(),reltol=1e-8,abstol=1e-8)` but all solver options from the [OrdinaryDiffEq.jl](https://github.com/SciML/OrdinaryDiffEq.jl) package are available.
-- After this, `∫!` is also equiped to postprocess results recorded into 🔴 via the 🔧 workflow, and the last step in `∫!` consiste in updating 📌 to be ready for continuing in a subsequent call to `∫!`.
+- After this, `∫!` is also equipped to postprocess results recorded into 🔴 via the 🔧 workflow, and the last step in `∫!` consists in updating 📌 to be ready for continuing in a subsequent call to `∫!`.
 """
 function ∫!(𝐼::Individuals,𝑇::Tuple)
     @unpack 🚄,📌,𝑃, 🔧, 🆔, 🔴, ∫ = 𝐼
