### 2021-02-04 07325cb
[test] comment out runs in docs/make.jl
Former-commit-id: 363cde22a8004844ef79ac87aeb20d3befc196d9
