### 2021-02-05 7b17df6
speed up docs build -- was hanging due to plots, especially gifs
Former-commit-id: c4df7103024a13c2d1c55792b52e7cb4349091fc [formerly 389955e9618899640413b7884ecef8c01c7a1121]
Former-commit-id: ca0725f24d7a54754ba277b5b50c181acd8752de
Former-commit-id: 048512531a3eec601ddcf29dfee6adfbb9770d22
