### 2021-02-05 c4df710
speed up docs build -- was hanging due to plots, especially gifs
Former-commit-id: 389955e9618899640413b7884ecef8c01c7a1121
