### 2021-02-05 cb0ab11
add  global_ocean_circulation jldoctest back
Former-commit-id: 92bb1f6298fbd8e1ec0441026409e10ba7a31cac [formerly 994c9f12e02da80c95989a30447e7b3502304104]
Former-commit-id: e1f27a0c25d11c734356d22ad432d637ea61fcc8
Former-commit-id: 48580eb25026df255e1d583610ee86db9f035c0b
