### 2021-02-05 JuliaClimate/IndividualDisplacements.jl#38 @gaelforget
> I'm running into the same issue wanting to apply this on a simple example grid of my own. I think the suggested syntax by @milankl could help here.
>
> It's great to have many examples in the docs, but for me it seems the first tutorial (single particle) is too complicated. It would be best if it can be self explanatory, with no code being included from separate files. How about something super simple like this?
>
> ```julia
> using Plots
>
> X = range(0, 2π, length=60)
> Y = range(0, π, length=30)
> f(x, y) = sin(x) + cos(y)
> ϕ = [f(x, y) for y in Y, x in X]
> u = [cos(x) for y in Y, x in X]
> v = [sin(y) for y in Y, x in X]
>
> heatmap(X, Y, ϕ)
> ```
>
> ![image](https://user-images.githubusercontent.com/4471859/102698106-33883400-423b-11eb-99b6-9cf6c7900fec.png)
>
> This would be a static 2D example. A next step could be adding a phase to show how this would work over time.

Hi @visr

Fair enough.

I have now revamped the second example (random flow field) using the suggested streamfunction as I understood it (below w. a tunable factor added) and promoted it to appear as the first example in the docs.

```
fac=0.1
f(x, y) = sin(x) + cos(y) #streamfunction
ϕ = fac*[f(x, y) for x in XC,y in YC] #streamfunction
uC = -fac*[sin(y) for x in XC,y in YC] #dphi/dy at cell center
vC = -fac*[cos(x) for x in XC,y in YC]; #-dphi/dx at cell center
```

This simplicity saves the need for calling an external function like `random_flow_field` or `solid_body_rotation` to define a more complicated flow field. Other flow field definition functions used the examples have been gathered in `examples/flow_fields.jl` (also aiming to make things more clear).

The data structure constructor calls was also inlined in `random_flow_field.jl` so that every function the user calls in the example is readily inlined or obtained from the package module (constructors). This degree of simplification was enabled by the code modifications done in version `0.2.7` to `0.2.9`.

A couple more aspects that I should mention about the revamped `examples/basics/random_flow_field.jl`
- the staggering of variables for the C-grid and grid scale and normalization to grid point units (which were not done in your above comment, or indeed sufficiently explained in the docs) are now also directly embedded in the now first example found in the docs
- this first example now uses plain Array rather than MeshArray by default; the optional code though shows how to switch between Array and MeshArray

The syntax is a bit different from the one suggested by @milankl but I am hoping that `examples/basics/random_flow_field.jl` (now the first example in the docs) achieves the intended goal.

Does this work?
