### 2021-02-05 JuliaClimate/IndividualDisplacements.jl#38 @gaelforget
> Hey Gaël, going through the documentation, I find in most examples that you've hidden the explanation for how to set up the NamedTuple Parameter 𝑃 which contains field like `u0`,`u1`,`v0`,`v1` etc. in some underlying function like `𝑃,Γ=example2_setup()`. I'd assume that most users will have their velocity fields in some form, e.g. simply as julia array loaded from netcdf, so I'd be great to guide them how they can pass these arrays (which may be defined on a C-grid etc.) into IndividualDisplacements.
>
> Yes, you mention MeshArrays.jl, but unfortunately reading that documentation I still don't know how I can, for example, take data from ShallowWaters.jl and compute the particle trajectories. My idea was to test IndividualDisplacements against the semi-Lagrangian tracer advection scheme therein.

Hi @milankl

Fair enough.

Please refer to https://github.com/JuliaClimate/IndividualDisplacements.jl/issues/38#issuecomment-774256677 for additional detail as well as the updated docs about constructors.

Let me just get to the bottom line first if that's ok -- the application to `ShallowWaters.jl` now reduces to something like:

```
using ShallowWaters
Prog = RunModel(Float32,Ndays=10,g=10,H=500,Fx0=0.12,output=true)

u=Prog.u
v=Prog.v
η=Prog.η
𝑇=[0.0,864000.0]
dx=20000.0

using IndividualDisplacements

𝐹=𝐹_Array2D(u/dx,u/dx,v/dx,v/dx,𝑇)

(np,nq)=size(u)
x=np*(0.4 .+ 0.2*rand(100))
y=nq*(0.4 .+ 0.2*rand(100));

𝐼=Individuals(𝐹,x,y)

∫!(𝐼)
```

and for plotting, e.g.:

```
p=dirname(pathof(IndividualDisplacements))
include(joinpath(p,"../examples/recipes_plots.jl"));

🔴_by_t = groupby(𝐼.🔴, :t)
anim = @animate for t in eachindex(🔴_by_t)
   phi_scatter(η,🔴_by_t[t])
end

pth=tempdir()*"/"
gif(anim, pth*"ShallowWaters.gif", fps = 10)
```

The normalization by dx is as documented in the first example found in the docs (`examples/basics/random_flow_field.jl`) and temporal interpolation between velocity fields is readily provided by the `FlowFields` constructors + e.g. NetCDF.jl to read output from ShallowWaters.jl

```
u=ncread("run0000/u.nc","u")[:,:,end]
v=ncread("run0000/v.nc","v")[:,:,end]
η=ncread("run0000/eta.nc","eta")[:,:,end]
t=ncread("run0000/u.nc","t")
𝑇=Float64.(t[end-1:end])
```

Does this work?
