### 2021-02-05 JuliaClimate/IndividualDisplacements.jl#38 @gaelforget
> Then I believe that some functionality as I already mentioned
>
> > ```julia
> > uv = MeshArray(u,v,x,y,grid-type=:C,geometry=:rectangle)
> > I = Individuals(📌,🔴,uv)
> > ```
>
> would greatly increase the usability of this package. I guess ideally a user would want to have functions for three steps
>
> * Provide velocity fields to IndividualDisplacements such they can be correctly interpreted
> * Provide particle initial conditions
> * Calculate trajectories
>
> I see that you have the `∫!` and the `Individuals` functions for the last two steps, but then I don't understand why in the particle set tutorial you still pass the initial coordinates through the `setup_point_cloud` function?

Thanks for the feedback and suggestions.

I have now streamlined step 1 further so that in rectangular domain case (the most common one maybe) it amounts to calling the corresponding `FlowFields` constructor for plain Arrays (`𝐹_Array2D`). Interpolation from grid-cell center to C-grid velocity points (at grid cell edges) and normalization by grid scale factors are simple Array operations in this case -- as documented in e.g. `random_flow_field.jl` now.

The more general case of global C-grids commonly used in climate models is when something like `MeshArrays.jl` can greatly facilitate gridded data manipulations which new users often struggle with in my experience. I tried to document these aspects as effectively as possible via the examples, but feel that moving some of the data munging steps to separate file (e.g. flow_fields.jl) improves different levels of reading of the docs.

So my hope is that "How to provide C-grid u,v?" is now sufficiently well addressed in the examples / docs and that this issue can therefore be closed.

Would you agree?

ps. At this point (incl. based on recent developments discussed in this thread) I believe that `IndividualDisplacements` will provide a solid basis for even more generalization in the future. In fact, one may hope that extensions to other grid topologies (or `grid-type` / `geometry` in your words) via `MeshArrays.jl` (or other packages) could occur relatively soon. But, I would also say that this goes beyond the current, stated, scope of `IndividualDisplacements.jl`
