### 2021-02-05 #2813 @gaelforget
@kthyng

Another update to let you know that I feel I have now completed the code and documentation updates (as of v0.2.9) needed to address https://github.com/JuliaClimate/IndividualDisplacements.jl/issues/38 where I just posted more detailed response to @milankl and @visr
