### 2021-02-06 392ee33
move set_up_P and update_P to examples
