+++ src/data_wrangling.jl
@@ -126,85 +126,6 @@ function postprocess_xy(sol,𝑃::FlowFields; id=missing, 𝑇=missing)
 end
 
 """
-    set_up_𝑃(k::Int,t::Float64,Γ::Dict,pth::String)
-
-Define the `𝑃` _parameter_ tuple given grid variables `Γ`, vertical level
-choice `k`, time `t` in `seconds`, and velocity fields obtained from
-files in `pth`.
-
-The two climatological months (`m0`,`m1`) that bracket time `t` are
-read to memory (e.g. months 12 & 1 then 1 & 2 and so on).
-
-_Note: the initial implementation approximates every month duration
-to 365 days / 12 months for simplicity._
-"""
-function set_up_𝑃(k::Int,t::Float64,Γ::Dict,pth::String)
-    XC=exchange(Γ["XC"]) #add 1 lon point at each edge
-    YC=exchange(Γ["YC"]) #add 1 lat point at each edge
-    iDXC=1. ./Γ["DXC"]
-    iDYC=1. ./Γ["DYC"]
-    γ=Γ["XC"].grid
-    mon=86400.0*365.0/12.0
-    func=Γ["update_location!"]
-    
-    𝐷 = (🔄 = update_𝑃!, pth=pth,
-         XC=XC, YC=YC, iDXC=iDXC, iDYC=iDYC,
-         k=k, msk=Γ["hFacC"][:, k])
-
-    tmp = dict_to_nt(IndividualDisplacements.NeighborTileIndices_cs(Γ))
-    𝐷 = merge(𝐷 , tmp)
-
-    𝑃=𝐹_MeshArray2D{Float64}(MeshArray(γ,Float64),MeshArray(γ,Float64),
-    MeshArray(γ,Float64),MeshArray(γ,Float64),[-mon/2,mon/2],func)
-
-    𝐷.🔄(𝑃,𝐷,0.0)
-
-    return 𝑃,𝐷
-end
-
-"""
-    update_𝑃!(𝑃::Union{NamedTuple,FlowFields},t::Float64)
-
-Update input data (velocity arrays) and time period (array) inside 𝑃 (𝑃.u0[:], etc, and 𝑃.𝑇[:])
-based on the chosen time `t` (in `seconds`). 
-
-_Note: for now, it is assumed that (1) input 𝑃.𝑇 is used to infer `dt` between consecutive velocity fields,
-(2) periodicity of 12 monthly records, (3) vertical 𝑃.k is selected -- but this could easily be generalized._ 
-"""
-function update_𝑃!(𝑃::Union{NamedTuple,FlowFields},𝐷::NamedTuple,t::Float64)
-    dt=𝑃.𝑇[2]-𝑃.𝑇[1]
-
-    m0=Int(floor((t+dt/2.0)/dt))
-    m1=m0+1
-    t0=m0*dt-dt/2.0
-    t1=m1*dt-dt/2.0
-
-    m0=mod(m0,12)
-    m0==0 ? m0=12 : nothing
-    m1=mod(m1,12)
-    m1==0 ? m1=12 : nothing
-
-    (U,V)=read_velocities(𝑃.u0.grid,m0,𝐷.pth)
-    u0=U[:,𝐷.k]; v0=V[:,𝐷.k]
-    u0[findall(isnan.(u0))]=0.0; v0[findall(isnan.(v0))]=0.0 #mask with 0s rather than NaNs
-    u0=u0.*𝐷.iDXC; v0=v0.*𝐷.iDYC; #normalize to grid units
[…]
