### 2021-02-06 8cbac0c
add nearest_to_xy; rm nitialize_*; update accordingly
