### 2021-02-06 b3b4aa7
add nearest_to_xy; rm nitialize_*; update accordingly
Former-commit-id: 8cbac0c7e977388661f7d9f404a274c33fa6e8ab
