### 2021-02-07 1074cb1
increase code coverage + fixes
