+++ examples/worldwide/three_dimensional_ocean.jl
@@ -146,3 +146,8 @@ set_up_individuals(𝐼::Individuals; nf=10000) = set_up_individuals(𝑃,Γ,�
 # p=PlotMakie(𝐼.🔴,100,180.);
 # display(p)
 # ```
+
+# ## 3.3 Alternatives (optional / unit testing)
+
+(x,y,z,f)=𝐼.📌[1]
+𝐽=Individuals(𝐼.𝑃,x,y,z,f)
