+++ test/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
+CyclicArrays = "6023044f-b3ef-4c22-8399-3ec90a0c9b75"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 MITgcmTools = "62725fbc-3a66-4df3-9000-e33e85b3a198"
