### 2021-02-07 3e4b963
improve docs section 2 and 3
Former-commit-id: 295dad7ceb45cb05c12b39198854739840078ee3 [formerly 95bac3320552a30083953d9727ecf912767c833e]
Former-commit-id: d7d2143e2977de16c9885e291836563fd6776a35
Former-commit-id: 915ee22a2f40655984dc0199de1168a03b7430c4
