+++ docs/src/workflow.md
@@ -1,23 +1,26 @@
-## Scope / Goals
-
-A central goal of this package is to support scientific analysis of climate model output and observed drifts of e.g. plastics in the Ocean or chemical coumpounds in the Atmosphere. 
-
-To start, the package supports all types of gridded model output from the [MIT General Circulation Model](https://mitgcm.readthedocs.io/en/latest/?badge=latest) via the [MeshArrays.jl](https://github.com/JuliaClimate/MeshArrays.jl) package ([docs found here](https://juliaclimate.github.io/MeshArrays.jl/dev/)) and ingestion of trajectory data which have been collected by the [Ocean Drifting Buoy Program](https://doi.org/10.1002/2016JC011716) ([movie](https://youtu.be/82HPnYBtoVo)).
-
 
 ## Typical Workflow
 
 As documented in the **examples**, the typical worflow is:
 
 1. set up `FlowFields` data structure
-1. set up `Individuals` with initial position 📌
-1. displace `Individuals` according to `FlowFields` (via	∫! and 🚄)
-1. post-process and record results (via 🔧 and 🔴)
-1. go back to `2` and continue if needed
+1. set up `Individuals` with initial position `📌`
+1. displace `Individuals` according to `FlowFields` (via	`∫!` and `🚄`)
+1. post-process and record results (via `🔧` and `🔴`)
+1. go back to `step 2` and continue if needed
 
 Steps `1` and `2` are done via data structures documented below. Both steps `3` and step `4` often take place within the call to `∫!`which can readily post-process results via 🔧 before recording them in 🔴 and finally updating positions 📌. Since 🔴 is in the [DataFrames](https://juliadata.github.io/DataFrames.jl/latest/) tabular format, it is easily manipulated, plotted, or saved after the fact (step `4` per se).
 
-**The examples** document simple methods to ingest time varying flow fields, three-dimensional ocean simulations, process oriented configurations, plotting tools, and data formats. For an overview of the examples, please refer to the **example guide**. The rest of this section is focused on the package's **data structures** and **core functions**.
+## Scope And Examples
+
+A central goal of this package is to support scientific analysis of climate model output and observed drifts of e.g. plastics in the Ocean or chemical coumpounds in the Atmosphere. 
+As an initial step, the package supports all types of gridded model output (on Arakawa C-grids) from the [MIT General Circulation Model](https://mitgcm.readthedocs.io/en/latest/?badge=latest) via the [MeshArrays.jl](https://github.com/JuliaClimate/MeshArrays.jl) package ([docs found here](https://juliaclimate.github.io/MeshArrays.jl/dev/)). 
+
+By convention, `IndividualDisplacements.jl` expects input flow fields to be provided in a uniform fashion: (1) normalized to grid index units (i.e. in 1/s rather than m/s units); (2) positive towards increasing indices; (3) using the Arakawa C-grid, with `u` (resp `v`) staggered by `-0.5` point in direction `1` (resp `2`) from grid cell centers. 
+
+**The examples** document simple methods to prepare and ingest such flow fields (time varying or not; in 2D or 3D) and derive individual displacements / trajectories from them. They cover simple grids often used in process studies, Global Ocean simulations normally done on more complex grids, plotting tools, and data formats. 
+
+For an overview of the examples, please refer to the **example guide**. The rest of this section is focused on the package's **data structures** and **core functions**.
 
 ## Data Structures
 
@@ -42,4 +45,5 @@ The velocity interpolation functions (🚄) carry out the central computation of
 - Preprocessing and postprocessing methods.
 - I/O routines to read (write) results from (to) file.
 
[…]
