### 2021-02-08 6d168ca
add great circle distance computation
Former-commit-id: 73bb05fae1a4483ea2d5a3576d2ab6d95edcf73f [formerly ecbe7cf90c3580e919c35fa2c2dfadd01a541c9c]
Former-commit-id: e77d61616cc16748b786c817c1ccef6d977d1c34
Former-commit-id: 4e5b2fd42523fa5975c421d94cb59fd1004686e7
