### 2021-02-08 73bb05f
add great circle distance computation
Former-commit-id: ecbe7cf90c3580e919c35fa2c2dfadd01a541c9c
