+++ src/IndividualDisplacements.jl
@@ -20,7 +20,7 @@ export FlowFields, convert_to_FlowFields
 export 𝐹_Array3D, 𝐹_Array2D, 𝐹_MeshArray3D, 𝐹_MeshArray2D
 export dxy_dt!, dxy_dt, dxyz_dt!, dxyz_dt, dxy_dt_CyclicArray, dxy_dt_replay
 export postprocess_MeshArray, add_lonlat!, postprocess_xy, interp_to_xy
-export nearest_to_xy, randn_lonlat, interp_to_lonlat
+export nearest_to_xy, randn_lonlat, interp_to_lonlat, gcdist
 export read_drifters, read_mds, read_velocities
 
 end # module
