### 2021-02-11 e7c6f24
update + streamline summary section after code changes (up to v0.2.10)
Former-commit-id: 8892534b54aaee98c8bed2a20990c9ae826f7132
