+++ docs/joss/paper.md
@@ -30,17 +30,18 @@ bibliography: paper.bib
 
 # Summary
 
-`IndividualDisplacements.jl` [@IndividualDisplacementsAugust2020] is focused on computation and analysis of individual displacements across the global climate system. _Individuals_ is used here as a generic term to represent points, particles, parcels, materials, etc that tend to be carried around the Earth by geophysical fluids (via oceanic currents or atmospheric flows for example). 
+`IndividualDisplacements.jl` [@IndividualDisplacementsAugust2020] is focused on computation and analysis of individual displacements across the global climate system. _Individuals_ is used here as a generic term to represent points, particles, parcels, materials, etc that tend to be carried around the Earth by geophysical fluids (via oceanic currents or atmospheric flows for example). Their _displacements_, or trajectories, are computed by deriving pointwise velocities from flow fields provided by the user and integrating over time.
 
-This package is aimed at the analysis of numerical models that simulate Oceanic and / or Atmospheric transport processes on, typically, global staggered C-grids [e.g. see @gmd-8-3071-2015]. Inter-operability with these climate model grids, based on `MeshArrays.jl` which was introduced at [JuliaCon 2018 (video link)](https://youtu.be/RDxAy_zSUvg) [@MeshArraysAugust2020], is a key feature. The chosen approach covers the comparatively simple grids used in e.g. satellite data products as well.
 
-`IndividualDisplacements.jl` is readily suited to exploit climate model output and other gridded data sets in research projects that involve tracking plankton communities, heat storage, or plastic garbage patches within the Ocean for example; dust, water, or chemical compounds within the Atmosphere, etc. It currently employs `OrdinaryDiffEq.jl` to integrate particle motions over time, `NetCDF.jl` and `CSV.jl` for I/O, and `DataFrames.jl` for diagnostics computed along particle trajectories.
+This Julia package is notably aimed at the analysis of numerical models that simulate Oceanic and / or Atmospheric transport processes on staggered C-grids -- those most often used in both regional and global modeling. Inter-operability with global climate model grids like the cube-sphere and lat-lon-cap grid [e.g. see @gmd-8-3071-2015] is a key feature enabled by `MeshArrays.jl` ([JuliaCon 2018 (video link)](https://youtu.be/RDxAy_zSUvg) ; [@MeshArraysAugust2020]). The chosen approach also readily supports simpler grids (cartesian, spherical, curvilinear grids) often used in process-oriented models, regional models, or satellite data products.
 
-The package is also intended to facilitate research involving, for example, model-data comparison, model intercomparison, data assimilation, state estimation, or machine learning by providing basic interfaces to observed displacement data sets (initially those collected by the Global Ocean Drifter Program and Argo array of drifting buoys) and to simulated displacement data sets generated internally, _online_, by climate models, or via other Lagrangian toolboxes (initially those from the MIT general circulation model).
+`IndividualDisplacements.jl` is thus readily suited to exploit climate model output and other gridded data sets in research projects that involve, for example, tracking plankton communities, heat storage, or plastic garbage patches within the Ocean; or dust, water, or chemical compounds within the Atmosphere. To achieve generality and interpretability, `IndividualDisplacements.jl` first defines two data structures (`FlowFields` and `Individuals`) that allow for simple, flexible user specifications of flow fields, initial individual positions, etc. It then adds a high-level API such that integrating trajectories for individuals `𝐼` amounts to a single `∫!(𝐼)` function call.
 
[…]
