+++ docs/joss/paper.bib
@@ -117,6 +117,18 @@ year = {2015}
   url          = {https://doi.org/10.5281/zenodo.2653669}
 }
 
+@software{FltGlobalOceanWorkflow2020,
+  author       = {Gael Forget},
+  title        = {{gaelforget/ECCOv4\_flt\_offline: initial public 
+                   release}},
+  month        = nov,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.2},
+  doi          = {10.5281/zenodo.4193582},
+  url          = {https://doi.org/10.5281/zenodo.4193582}
+}
+
 @article{Rousselet2020,
 author = {Rousselet, Louise and Cessi, Paola and Forget, Gael},
 title = {Routes of the Upper Branch of the Atlantic Meridional Overturning Circulation according to an Ocean State Estimate},
@@ -189,3 +201,15 @@ year = {2020}
   doi          = {10.5281/zenodo.3992421},
   url          = {https://doi.org/10.5281/zenodo.3992421}
 }
+
+@Article{gmd-2020-385,
+AUTHOR = {Peytavin, A. and Sainte-Rose, B. and Forget, G. and Campin, J.-M.},
+TITLE = {Ocean Plastic Assimilator v0.1: Assimilation of Plastics Concentration Data Into Lagrangian Dispersion Models},
+JOURNAL = {Geoscientific Model Development Discussions},
+VOLUME = {2021},
+YEAR = {2021},
+PAGES = {1--17},
+URL = {https://gmd.copernicus.org/preprints/gmd-2020-385/},
+DOI = {10.5194/gmd-2020-385}
+}
+
