### 2021-02-11 95f848a
expand on context v other packages + references
Former-commit-id: 59d6a7c78ae46f59b6937ee1f8d5e0cdf88fd691
