+++ docs/joss/paper.md
@@ -43,17 +43,17 @@ Internally, the package currently employs `OrdinaryDiffEq.jl` to integrate parti
 
 The examples folder, which is unit tested upon building the documentation online for each code revision, already covers several common grid cases, two-dimensional and three-dimensional flows, steady-state and time-variable flows. It demonstrates interpolation and diagnostics methods, three plotting libraries, and two notebook systems. Minor parameter changes in these examples should suffice to start applying `IndividualDisplacements.jl` to any other MITgcm simulation of e.g. atmospheric or oceanic turbulence [@MITgcm2020].
 
-# Statement of need 
+# Statement of need
 
-Lagrangian simulation and analysis frameworks such as `IndividualDisplacements.jl` have been widely used across scientific domains involved with geophysical fluids for decades [see @FLEXPART],[@VanSebille2018 for recent reviews in oceanography or atmospheric sciences]. 
+Lagrangian simulation and analysis frameworks such as `IndividualDisplacements.jl` have been widely used across scientific domains involved with geophysical fluids for decades [see @FLEXPART],[@VanSebille2018 for recent reviews in oceanography or atmospheric sciences]. `IndividualDisplacements.jl` is most directly related to the `MITgcm/flt` fortran package which the author recently extended [@MITgcm2020]. This new Julia package was in fact partly motivated by a need to provide an easier and simpler alternative to using `MITgcm/flt` in its offline mode in order to handle global grids not readily supported by other packages [e.g. @gmd-8-3071-2015, @FltGlobalOceanWorkflow2020, @Rousselet2020].
 
-`IndividualDisplacements.jl` provides a bridge between this vast community of domain experts, who are typically used to older languages like Fortran, C, Matlab, or Python, and the rapidly growing Julia community and package ecosystem. Julia's native parallelism, GPU support, custom array types, differential programming tools, and plotting libraries indeed all offer great new opportunities for exploiting state of the art HPC climate model simulations and the large data sets that they routinely generate.
+`IndividualDisplacements.jl` notably aims to provide a bridge between the vast community of domain experts, who are typically used to older languages like Fortran, C, Matlab, or Python, and the rapidly growing Julia community and package ecosystem. Julia's native parallelism, GPU support, custom array types, differential programming tools, and plotting libraries indeed all offer great new opportunities for exploiting state of the art HPC climate model simulations and the large data sets that they routinely generate.
 
 The development of `IndividualDisplacements.jl` was motivated not just by our scientific needs and ongoing research projects (see below) but also by the need for more effective tools for climate education, citizen science, and advocacy. In the documentation and examples folder, this is highlighted by the use of unicode in the API, `Pluto.jl` notebooks, and `Makie.jl` animations, which showcase Julia's expressivity and reactivity. 
 
 `IndividualDisplacements.jl` readily supports all common `MITgcm` grids and configurations via `MeshArrays.jl` (incl. Ocean,  Atmosphere, sea-ice, bio-geo-chemistry, and ecology). This, by itself, yields a large pool of potential scientific applications and expected users -- including via the many ongoing research projects that rely on adjoint-optimized, data-constrained solutions from the OCCA/ECCO project [see @Forget2010][@gmd-8-3071-2015]. 
 
-At this stage, `IndividualDisplacements.jl` is considered production-ready, such that we could immediately start transferring several ongoing research collaborations that use MITgcm and ECCO [e.g. @Rousselet2020], [@Forget2019] from fortran to Julia. Extension to other models may start with MERRA2 / MITgcm coupled model runs later this year [@Strobach2020]. 
[…]
