+++ docs/joss/paper.md
@@ -35,9 +35,9 @@ bibliography: paper.bib
 
 This Julia package is notably aimed at the analysis of numerical models that simulate Oceanic and / or Atmospheric transport processes on staggered C-grids -- those most often used in both regional and global modeling. Inter-operability with global climate model grids like the cube-sphere and lat-lon-cap grid [e.g. see @gmd-8-3071-2015] is a key feature enabled by `MeshArrays.jl` ([JuliaCon 2018 (video link)](https://youtu.be/RDxAy_zSUvg) ; [@MeshArraysAugust2020]). The chosen approach also readily supports simpler grids (cartesian, spherical, curvilinear grids) often used in process-oriented models, regional models, or satellite data products.
 
-`IndividualDisplacements.jl` is thus readily suited to exploit climate model output and other gridded data sets in research projects that involve, for example, tracking plankton communities, heat storage, or plastic garbage patches within the Ocean; or dust, water, or chemical compounds within the Atmosphere. To achieve generality and interpretability, `IndividualDisplacements.jl` first defines two data structures (`FlowFields` and `Individuals`) that allow for simple, flexible user specifications of flow fields, initial individual positions, etc. It then adds a high-level API such that integrating trajectories for individuals `𝐼` amounts to a single `∫!(𝐼)` function call.
+`IndividualDisplacements.jl` is thus readily suited to exploit climate model output and other gridded data sets in research projects that involve, for example, tracking plankton communities, heat storage, or plastic garbage patches within the Ocean; or dust, water, or chemical compounds within the Atmosphere. To achieve generality and interpretability, `IndividualDisplacements.jl` first defines two data structures (`FlowFields` and `Individuals`) that allow for simple, flexible user specifications of flow fields, initial individual positions, etc. It then adds a high-level API such that integrating trajectories for individuals `𝐼` amounts to a single function call, $\int!\,(\,𝐼\,)$.
 
-Internally, the package currently employs `OrdinaryDiffEq.jl` to integrate particle motions over time (within the `∫!(𝐼)` function call), `NetCDF.jl` and `CSV.jl` for I/O, and `DataFrames.jl` for diagnostics computed along particle trajectories (within `∫!(𝐼)` or afterwards). The initial test suite [see @IndividualDisplacementsAugust2020] is based not only on idealized flow fields and toy-models, but also on data-constrained Ocean simulations from the OCCA/ECCO projects [@Forget2010],[@gmd-8-3071-2015], [@Forget2018setup], which are also used in marine ecosystem simulations in CBIOMES [@cbiomes2019]. The OCCA/ECCO gridded ocean circulation estimates are retrieved from a permanent dataverse archive [@OCCAdataverse], [@Forget2016dataverse] via the Julia Artifacts system.
+Internally, the package currently employs `OrdinaryDiffEq.jl` to integrate particle motions over time (within the $\int!\,(\,𝐼\,)$ function call), `NetCDF.jl` and `CSV.jl` for I/O, and `DataFrames.jl` for diagnostics computed along particle trajectories (within $\int!\,(\,𝐼\,)$ or afterwards). The initial test suite [see @IndividualDisplacementsAugust2020] is based not only on idealized flow fields and toy-models, but also on data-constrained Ocean simulations from the OCCA/ECCO projects [@Forget2010],[@gmd-8-3071-2015], [@Forget2018setup], which are also used in marine ecosystem simulations in CBIOMES [@cbiomes2019]. The OCCA/ECCO gridded ocean circulation estimates are retrieved from a permanent dataverse archive [@OCCAdataverse], [@Forget2016dataverse] via the Julia Artifacts system.
 
[…]
