### 2021-02-11 d2397bd
replace unicode (not rendered) with tex integral formula
Former-commit-id: d51d7e00d58a14acdbd72e798cb73deaafa5aa6a [formerly 05285b5ce0aad245cfb0e13f7ec36e4cccd24430]
Former-commit-id: c113ab9cc97b25deafb19a453e1d7844986367b9
Former-commit-id: 33131b10226aae6a6c1a5ca2e74142a1015a654c
