### 2021-02-11 3b5a313
minor fixes and doi updates
Former-commit-id: 857d58ca416301c25523e0fe43dfe6db21962fbf [formerly 9c1dcb1873b5d1430d16e3bafc974594b614f4c6]
Former-commit-id: 71356ae6ca0a4ebd084c7f3f25faff4b4cdb77e3
Former-commit-id: 05cd23095453005a524f61bf84208d922a469bf1
