+++ docs/joss/paper.bib
@@ -191,15 +191,15 @@ year = {2020}
   url          = {https://doi.org/10.5281/zenodo.3995076}
 }
 
-@software{IndividualDisplacementsAugust2020,
+@software{IndividualDisplacementsFeb2021,
   author       = {Gael Forget},
-  title        = {JuliaClimate/IndividualDisplacements.jl: v0.1.15},
-  month        = aug,
-  year         = 2020,
+  title        = {JuliaClimate/IndividualDisplacements.jl: v0.2.10},
+  month        = feb,
+  year         = 2021,
   publisher    = {Zenodo},
-  version      = {v0.1.15},
-  doi          = {10.5281/zenodo.3992421},
-  url          = {https://doi.org/10.5281/zenodo.3992421}
+  version      = {v0.2.10},
+  doi          = {10.5281/zenodo.4517441},
+  url          = {https://doi.org/10.5281/zenodo.4517441}
 }
 
 @Article{gmd-2020-385,
