### 2021-02-11 #2813 @gaelforget
> @gaelforget Ok thanks for the update. I'll wait to hear from @milankl and @visr about their status after this work.

I just posted an updated version of the paper itself (in the `josspaper_rev01` branch) to reflect code changes that have happened since original submission, and hopefully clarify the functionality summary.

I also expended a bit on the context vs other particle tracking software as advised during the pre-review at the top of section 2 (please note that the included reference from Van Sebille et al has a lot more on this) and updated references & funding aknowledgments.

Having verified that the updated paper.md compiles and renders properly, I feel that `josspaper_rev01` is now ready to merge into the original `josspaper` branch. However, I stopped short of merging the PR as I wanted to check in with you first:

Is it a good time to merge this PR https://github.com/JuliaClimate/IndividualDisplacements.jl/pull/61 and update the `josspaper` branch or should I wait on that?
