### 2021-02-18 8800ee8
add stproj, stproj_inv, reset_xy + fix init_global_randn, add_lonlat
Former-commit-id: 9cff793e13aae70ffb8e39a790d3837c87d77246 [formerly 93d72808242ff81fa4cbf297b951685f37e7469c]
Former-commit-id: 5f2dfb4ebd7c6bd84831cb1f1bb5b003e2e5e471
Former-commit-id: 00bf837872e09b4fa5848bd50928a09344c5b7ba
