### 2021-02-18 9cff793
add stproj, stproj_inv, reset_xy + fix init_global_randn, add_lonlat
Former-commit-id: 93d72808242ff81fa4cbf297b951685f37e7469c
