+++ src/data_wrangling.jl
@@ -67,26 +67,19 @@ end
 Add lon & lat to dataframe using "exchanged" XC, YC
 """
 function add_lonlat!(df::DataFrame,XC,YC)
-    x=df[!,:x];
-    y=df[!,:y];
-    f=Int.(df[!,:fid]);
-    dx,dy=(x - floor.(x) .+ 0.5,y - floor.(y) .+ 0.5);
-    i_c = Int32.(floor.(x)) .+ 1;
-    j_c = Int32.(floor.(y)) .+ 1;
+    x = cosd.(YC) * cosd.(XC)
+    y = cosd.(YC) * sind.(XC)
+    z = sind.(YC)
 
-    lon=zeros(length(x),4)
-    [lon[k,:]=XC[f[k]][i_c[k]:i_c[k]+1,j_c[k]:j_c[k]+1][:] for k in 1:length(i_c)]
-    lat=zeros(length(x),4)
-    [lat[k,:]=YC[f[k]][i_c[k]:i_c[k]+1,j_c[k]:j_c[k]+1][:] for k in 1:length(i_c)]
-
-    k=findall(vec(maximum(lon,dims=2)-minimum(lon,dims=2)) .> 180.0)
-    tmp=view(lon,k,:)
-    tmp[findall(tmp.<0.0)]=tmp[findall(tmp.<0.0)] .+ 360.0
-
-    df.lon=(1.0 .-dx).*(1.0 .-dy).*lon[:,1]+dx.*(1.0 .-dy).*lon[:,2] +
-         (1.0 .-dx).*dy.*lon[:,3]+dx.*dy.*lon[:,4]
-    df.lat=(1.0 .-dx).*(1.0 .-dy).*lat[:,1]+dx.*(1.0 .-dy).*lat[:,2] +
-         (1.0 .-dx).*dy.*lat[:,3]+dx.*dy.*lat[:,4]
+    df.lon=0*df.x
+    df.lat=0*df.x
+    for i in eachindex(df.lon)
+        xx=interp_to_xy(df.x[i],df.y[i],df.fid[i],x)
+        yy=interp_to_xy(df.x[i],df.y[i],df.fid[i],y)
+        zz=interp_to_xy(df.x[i],df.y[i],df.fid[i],z)
+        df.lat[i] = asind(zz/sqrt(xx^2+yy^2+zz^2))
+        df.lon[i] = atand(yy, xx)
+    end
 
     return df
 end
@@ -208,4 +201,82 @@ function interp_to_xy(df::DataFrame,Zin::MeshArray)
 
     return (1.0 .-dx).*(1.0 .-dy).*Z[:,1]+dx.*(1.0 .-dy).*Z[:,2] +
            (1.0 .-dx).*dy.*Z[:,3]+dx.*dy.*Z[:,4]
+end
+
+function interp_to_xy(x,y,f,Zin::MeshArray)
+    dx,dy=(x - floor(x) + 0.5,y - floor(y) + 0.5)
+    i_c = Int(floor(x)) + 1
+    j_c = Int(floor(y)) + 1
+    ff=Int(f)
+
+    return (1.0-dx)*(1.0-dy)*Zin[ff][i_c,j_c] +
+    dx*(1.0-dy)*Zin[ff][i_c+1,j_c] +
+    (1.0-dx)*dy*Zin[ff][i_c,j_c+1] +
+    dx*dy*Zin[ff][i_c+1,j_c+1]
+end
+
+"""
+    stproj(XC,YC;XC0=0.0,YC0=90.0)
+
+Apply to XC,YC (longitude, latitude) the stereographic projection
+which would put XC0,YC0 (longitude, latitude) at x,y=0,0
+"""
+function stproj(XC,YC,XC0=0.0,YC0=90.0)
+
+# compute spherical coordinates:
[…]
