### 2021-02-19 1206913
streamline global / ECCO config
Former-commit-id: 38a00532b2d581572e7916058e557d2fc89067c8
