### 2021-02-19 38a0053
streamline global / ECCO config
