+++ src/API.jl
@@ -201,7 +201,7 @@ function Individuals(𝐹::𝐹_MeshArray3D,x,y,z,fid)
     🔴 = DataFrame(ID=Int[], x=Float64[], y=Float64[], z=Float64[], fid=Int64[], t=Float64[])
     function 🔧(sol,𝑄::FlowFields;id=missing,𝑇=missing)
         df=postprocess_MeshArray(sol,𝐹,id=id,𝑇=𝑇)
-        z=sol[3,:]
+        z=[sol[1,i,j][1] for i in 1:size(sol,2), j in 1:size(sol,3)]
         df.z=z[:]
         return df
     end
