+++ examples/helper_functions.jl
@@ -18,28 +18,17 @@ function init_global_randn(np ::Int , 𝑃::NamedTuple)
 end
 
 """
-    reset_lonlat!(𝐼::Individuals)
+    reset_📌!(𝐼::Individuals,frac::Number,📌::Array)
 
-Randomly select a fraction (𝐼.𝑃.frac) of the particles and reset their positions.
+Randomly select a fraction (frac) of the particles and reset 
+their positions (𝐼.📌) to a random subset of the specificed 📌.
 """
-function reset_lonlat!(𝐼::Individuals,𝐷::NamedTuple)
-    np=length(𝐼.🆔)
-    n_reset = Int(round(𝐷.frac*np))
-    v0=init_global_randn(n_reset , 𝐷)
-    n_reset=min(n_reset,size(v0,2))
-    k_reset = rand(1:np, n_reset)
-    v0 = permutedims([v0[:,i] for i in 1:size(v0,2)])
-    𝐼.📌[k_reset].=v0[1:n_reset]
-    isempty(𝐼.🔴.ID) ? m=maximum(𝐼.🆔) : m=max(maximum(𝐼.🔴.ID),maximum(𝐼.🆔))
-    𝐼.🆔[k_reset]=collect(1:n_reset) .+ m
-end
-
-function reset_xy!(𝐼::Individuals,𝐷::NamedTuple)
+function reset_📌!(𝐼::Individuals,frac::Number,📌::Array)
     np=length(𝐼.🆔)
     n_reset = Int(round(𝐷.frac*np))
     k_reset = rand(1:np, n_reset)
     l_reset = rand(1:np, n_reset)
-    𝐼.📌[k_reset]=permutedims([xy[:,l_reset[i]] for i in 1:n_reset])
+    𝐼.📌[k_reset]=deepcopy(📌[l_reset])
     isempty(𝐼.🔴.ID) ? m=maximum(𝐼.🆔) : m=max(maximum(𝐼.🔴.ID),maximum(𝐼.🆔))
     𝐼.🆔[k_reset]=collect(1:n_reset) .+ m
 end
@@ -88,21 +77,31 @@ function set_up_FlowFields(k::Int,Γ::Dict,pth::String)
     mon=86400.0*365.0/12.0
     func=Γ["update_location!"]
     
+    if k==0
+        msk=Γ["hFacC"]
+        (_,nr)=size(msk)
+        𝑃=𝐹_MeshArray3D{Float64}(MeshArray(γ,Float64,nr),MeshArray(γ,Float64,nr),
+        MeshArray(γ,Float64,nr),MeshArray(γ,Float64,nr),
+        MeshArray(γ,Float64,nr+1),MeshArray(γ,Float64,nr+1),
+        [-mon/2,mon/2],func)
+    else
+        msk=Γ["hFacC"][:, k]
+        𝑃=𝐹_MeshArray2D{Float64}(MeshArray(γ,Float64),MeshArray(γ,Float64),
+        MeshArray(γ,Float64),MeshArray(γ,Float64),[-mon/2,mon/2],func)    
+    end
+    
     𝐷 = (🔄 = update_FlowFields!, pth=pth,
          XC=XC, YC=YC, iDXC=iDXC, iDYC=iDYC,
-         k=k, msk=Γ["hFacC"][:, k])
+         k=k, msk=msk, θ0=similar(msk), θ1=similar(msk))
 
[…]
