### 2021-02-19 599d8ee
allow constructor customization via NamedTuple
Former-commit-id: 65d4ff97adaaec9ead3aa853d2d18cb20eeef543
