### 2021-02-20 d80ed43
distinct update_location_llc, improve add_lonlat
