### 2021-02-21 ee63f6c
update basic Plots and Makie recipes
