### 2021-02-21 f26cef2
update basic Plots and Makie recipes
Former-commit-id: 29fc7c16a94a9e321bbfed0a59df3759184c87ba [formerly ee63f6c480e858383116a54d6f8ca8044b72c854]
Former-commit-id: 9a6edabfb7ac735c76bedb245c2214b5979285e1
Former-commit-id: 9d52b89ff640172ca7a24e3b6bb69fb58ea80384
