### 2021-02-23 #2813 @kthyng
@gaelforget it doesn't matter to me. If your goal is to wait to merge until the paper has absolutely no more changes, then you should wait until I do a final read through at the end of the review. Otherwise it doesn't really matter from the JOSS perspective.
