### 2021-02-23 #2813 @gaelforget
> @gaelforget it doesn't matter to me. If your goal is to wait to merge until the paper has absolutely no more changes, then you should wait until I do a final read through at the end of the review. Otherwise it doesn't really matter from the JOSS perspective.

Thanks for the clarification. My question came from being unsure whether I should expect detailed comments on the paper itself from reviewers and at which stage of the review process.

Accordingly, I just merged the `josspaper_rev01` into `josspaper` so everything on my end is in sync right now with the latest release (v0.2.11).

A couple follow up questions, @kthyng :

- could / should the pdf available to reviewers linked in https://github.com/openjournals/joss-reviews/issues/2813#issuecomment-723257060 be updated?
- I though about adding maybe a pair of plots illustrating results from the package (e.g. one ocean, one atmosphere); can I do that and merge in the josspaper paper branch either now or later ?
