### 2021-02-24 #2813 @kthyng
> Thanks for the clarification. My question came from being unsure whether I should expect detailed comments on the paper itself from reviewers and at which stage of the review process.

Oh sorry I misunderstood your question. The reviewers should go through your paper as part of the review process as they go through the checklist so you do want it clearly available.

> * could / should the pdf available to reviewers linked in [#2813 (comment)](https://github.com/openjournals/joss-reviews/issues/2813#issuecomment-723257060) be updated?

Yes, I can run this command now and you can run it whenever you want, too.

> * I though about adding maybe a pair of plots illustrating results from the package (e.g. one ocean, one atmosphere); can I do that and merge in the josspaper paper branch either now or later ?

I'd add it now so the reviewers can consider the paper how you intend it to be.
