### 2021-03-11 f92c210
first try at including plots
Former-commit-id: af86586e8b1e6fcbeff779aa4339a11c9ef12275
