### 2021-03-11 #2813 @gaelforget
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02813/joss.02813/10.21105.joss.02813.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.02813/joss.02813/10.21105.joss.02813.pdf) 📄 👈

I just added a couple illustrations of the package applied to Atmosphere (#1) and Ocean (#2), plus a reference to them at the end of paragraph 1
