### 2021-03-13 #2813 @milankl
Hey @gaelforget, nice work! Yes, so I managed to feed IndividualDisplacements.jl with data from ShallowWaters.jl and can verify the particle advection here with the semi-Lagrangian advection scheme therein. So that ticks all the boxes from above. I'm happy for this work to get published so that hopefully more people will start using it, improving its features. There's still some issues that I may raise in the repo, but they are beyond the scope of this review. Thanks @kthyng for your efforts too.

[![000200](https://user-images.githubusercontent.com/25530332/111016072-c43a6000-83a3-11eb-9da1-794a0468fd09.png)
Click to see the video on twitter](https://twitter.com/milankloewer/status/1370561010856251393?s=20)
