### 2021-03-13 JuliaClimate/IndividualDisplacements.jl#38 @milankl
Thanks @gaelforget for the work you've put in here. Yes, I think the docs show now nicely how this can be done. I'll close this issue for now.
