### 2021-03-25 2dd33dd
resolve case where 𝑇 is not Array{T}
Former-commit-id: c956130bd7af97c99cad72c5259644a3f40a74fc [formerly dc910995236a07e61dc675f64fa064fcf42dfb00]
Former-commit-id: ccd7833414610b86285571d7d40efcc42d4a906a
Former-commit-id: b55d9d9e94ab14da7d96f34c17f91bcdc87dda7e
