### 2021-03-25 c956130
resolve case where 𝑇 is not Array{T}
Former-commit-id: dc910995236a07e61dc675f64fa064fcf42dfb00
