### 2021-03-25 08a6844
improve FlowFields constructors (checkbounds + dispatch)
Former-commit-id: 791035c52c590a8cad3edc9a5a49f0a5199d693b [formerly 32b5e220157e899e15d74670376c79d00702a305]
Former-commit-id: a40e2948e77d91433ef8b869fe816cd8576e8afd
Former-commit-id: dbe7382d3cf381fa7b1a95e9fe1aa51ba0cab8fd
