+++ examples/basics/solid_body_rotation.jl
@@ -39,7 +39,7 @@ np,nz=16,4 #gridded domain size (horizontal and vertical)
 
 u,v,w=solid_body_rotation(np,nz) #staggered velocity arrays
 
-𝐹=𝐹_Array3D{eltype(u)}(u,u,v,v,0*w,1*w,[0,19.95*2*pi]); #FlowFields data structure
+𝐹=FlowFields(u,u,v,v,0*w,1*w,[0,19.95*2*pi]); #FlowFields data structure
 
 #nb # %% {"slideshow": {"slide_type": "slide"}, "cell_type": "markdown"}
 # ### 1.3 Initialize Individuals
