### 2021-03-25 7848236
update examples using revised FlowFields constructors
