### 2021-03-26 #2813 @kthyng
@gaelforget A few things in your paper to fix. Let me know if you have questions:

1. Please check all of your inline references. If they are not meant to be parenthetical, leave off the [] around the reference so they are not parenthetical. This comes up a lot in the paper.
2. Check how you refer to the languages and other items through the paper. "Fortran" for example is all lower case one place but I believe should be "Fortran".
3. Check your references in detail. One has "{ECCO}" in it and presumably shouldn't, your name sometimes has an umlaut but not always, and some words are not capitalized that should be like Lagrangian and Atlantic. You can preserve capitalization by adding {} around the words in the .bib file that should be kept as is. Please check for other issues, too.
