### 2021-03-26 #2813 @kthyng
@gaelforget also fix that invalid doi please
