### 2021-04-10 #2813 @gaelforget
> @gaelforget also fix that invalid doi please

- Should be all set now on the references, I think -- unless there is a way, which I did not find, to remove parentheses around the year in e.g. `(Gaël Forget (2010), G. Forget et al. (2015), Gaël Forget (2018))`
- `Fortran` is now capitalized throughout
- I changed a few cosmetic details (mostly moving the plots to page 2) following a final pass on the paper
