### 2021-04-10 #2813 @gaelforget
> @gaelforget What version do you want associated with this submission? Should it be updated?
>
> Also, please archive your software at a place like Zenodo and check back here with the doi they give you. Please modify the metadata of the archive so that the title and author list exactly match your JOSS submission.

Thanks for the very helpful directions

I would like version `v0.3.0`, which I released just before this post, to be associated with this submission.

The corresponding reference, with edited metadata, is

Gael Forget. (2021, April 10). IndividualDisplacements.jl : a Julia package to simulate and study particle displacements within the climate system (Version v0.3.0). Zenodo. http://doi.org/10.5281/zenodo.4678742
