### 2021-04-10 #2813 @gaelforget
> @gaelforget Ack, sorry. I have been experimenting with labeling my email and I was too effective in hiding away my JOSS emails. Oops!
>
> I see that both reviewers are satisfied with your submission, so I will go through some final steps now.

Hi @kthyng

Thank you very much for the guidelines and all that you have done on this submission. Please find above responses to the issues listed. Thanks again
