### 2021-04-14 #2813 @gaelforget
> Hi @gaelforget, this is nearly ready for acceptance, but the parenthetical issues with the references throughout the paper remain.
>
> When citing things, you do not need to include "e.g." or "see" or "for example"—this is implied whenever you cite one or more publications. But, this is more a preference, I suppose. The only real places that need to be fixed are when you have the nested parentheses like in the fourth paragraph; the appropriate syntax for grouped citations is `[@author1:2001; @author2:2001]`, which formats like -> "(Author1 et al., 2001; Author2 et al., 2002)".
>
> I've made these changes directly in [JuliaClimate/IndividualDisplacements.jl#71](https://github.com/JuliaClimate/IndividualDisplacements.jl/pull/71), can you merge this?

Merged. Thanks @kyleniemeyer
