### 2021-04-22 #2813 @gaelforget
> Hi @gaelforget, looks like there were still some issues, or perhaps I made mistakes: [JuliaClimate/IndividualDisplacements.jl#72](https://github.com/JuliaClimate/IndividualDisplacements.jl/pull/72)

Just did a couple more tweaks. Although I feel unsure what the desired rendering really is in some cases, this looks good to me now.
