### 2021-04-28 #2813 @gaelforget
> Congrats @gaelforget on your article's publication in JOSS!
>
> Many thanks to @milankl and @visr for reviewing this, and @kthyng for editing.

Awesome news! Thank you @kyleniemeyer

Thank you @milankl and @visr -- your reviews definitely helped improve IndividualDisplacements.jl

Thank you also @kthyng and @kyleniemeyer for editing
